### 2025-08-04 #8727 @editorialbot
[PRE REVIEW]: RustSASA: A Rust Crate for Accelerated Solvent Accessible Surface Area Calculations
**Submitting author:** @maxall41 (<a href="http://orcid.org/0000-0002-0959-1164">Maxwell James Campbell</a>)
**Repository:** https://github.com/maxall41/RustSASA
**Branch with paper.md** (empty if default branch):
**Version:** 0.3.1
