### 2025-08-05 83d2d57
Bump quick-xml from 0.38.0 to 0.38.1
Bumps [quick-xml](https://github.com/tafia/quick-xml) from 0.38.0 to 0.38.1.
- [Release notes](https://github.com/tafia/quick-xml/releases)
- [Changelog](https://github.com/tafia/quick-xml/blob/master/Changelog.md)
- [Commits](https://github.com/tafia/quick-xml/compare/v0.38.0...v0.38.1)

---
updated-dependencies:
- dependency-name: quick-xml
  dependency-version: 0.38.1
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
