+++ Cargo.toml
@@ -19,7 +19,7 @@ snafu = "0.8.6"
 clap = { version = "4.5.41", features = ["derive"] }
 serde = "1.0.219"
 serde_json = "1.0.140"
-quick-xml = { version = "0.38.0", features = ["serialize"] }
+quick-xml = { version = "0.38.1", features = ["serialize"] }
 indicatif = "0.18.0"
 mimalloc = { version = "0.1.47", default-features = false }
 
