### 2025-08-07 #8727 @maxall41
Could you provide an estimate of when the paper will be reviewed, or the position of the paper in the queue?
