+++ Cargo.toml
@@ -16,7 +16,7 @@ pdbtbx = { path = "pdbtbx", version = "0.12.0" }             # We have to use a 
 nalgebra = "0.33.2"
 rayon = "1.10.0"
 snafu = "0.8.6"
-clap = { version = "4.5.41", features = ["derive"] }
+clap = { version = "4.5.44", features = ["derive"] }
 serde = "1.0.219"
 serde_json = "1.0.140"
 quick-xml = { version = "0.38.0", features = ["serialize"] }
