### 2025-08-26 a66d7ee
Bump snafu from 0.8.6 to 0.8.7
Bumps [snafu](https://github.com/shepmaster/snafu) from 0.8.6 to 0.8.7.
- [Changelog](https://github.com/shepmaster/snafu/blob/main/CHANGELOG.md)
- [Commits](https://github.com/shepmaster/snafu/compare/0.8.6...0.8.7)

---
updated-dependencies:
- dependency-name: snafu
  dependency-version: 0.8.7
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
