### 2025-09-02 0d7a69c
Bump snafu from 0.8.7 to 0.8.8
Bumps [snafu](https://github.com/shepmaster/snafu) from 0.8.7 to 0.8.8.
- [Changelog](https://github.com/shepmaster/snafu/blob/main/CHANGELOG.md)
- [Commits](https://github.com/shepmaster/snafu/compare/0.8.7...0.8.8)

---
updated-dependencies:
- dependency-name: snafu
  dependency-version: 0.8.8
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
