### 2025-09-08 4e8724b
Bump snafu from 0.8.8 to 0.8.9
Bumps [snafu](https://github.com/shepmaster/snafu) from 0.8.8 to 0.8.9.
- [Changelog](https://github.com/shepmaster/snafu/blob/main/CHANGELOG.md)
- [Commits](https://github.com/shepmaster/snafu/compare/0.8.8...0.8.9)

---
updated-dependencies:
- dependency-name: snafu
  dependency-version: 0.8.9
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
