+++ Cargo.toml
@@ -15,7 +15,7 @@ rust-version = "1.85.0"
 pdbtbx = { path = "pdbtbx", version = "0.12.0" }             # We have to use a git module because you cannot use a git+rev version in a published crate for some reason...
 nalgebra = "0.33.2"
 rayon = "1.11.0"
-snafu = "0.8.8"
+snafu = "0.8.9"
 clap = { version = "4.5.46", features = ["derive"] }
 serde = "1.0.219"
 serde_json = "1.0.143"
