+++ Cargo.toml
@@ -18,7 +18,7 @@ rayon = "1.11.0"
 snafu = "0.8.8"
 clap = { version = "4.5.46", features = ["derive"] }
 serde = "1.0.219"
-serde_json = "1.0.143"
+serde_json = "1.0.145"
 quick-xml = { version = "0.38.3", features = ["serialize"] }
 indicatif = "0.18.0"
 mimalloc = { version = "0.1.48", default-features = false }
