+++ Cargo.toml
@@ -16,7 +16,7 @@ pdbtbx = { path = "pdbtbx", version = "0.12.0" }             # We have to use a 
 nalgebra = "0.33.2"
 rayon = "1.11.0"
 snafu = "0.8.8"
-clap = { version = "4.5.46", features = ["derive"] }
+clap = { version = "4.5.48", features = ["derive"] }
 serde = "1.0.219"
 serde_json = "1.0.143"
 quick-xml = { version = "0.38.3", features = ["serialize"] }
