+++ Cargo.toml
@@ -17,7 +17,7 @@ nalgebra = "0.33.2"
 rayon = "1.11.0"
 snafu = "0.8.8"
 clap = { version = "4.5.46", features = ["derive"] }
-serde = "1.0.219"
+serde = "1.0.228"
 serde_json = "1.0.143"
 quick-xml = { version = "0.38.3", features = ["serialize"] }
 indicatif = "0.18.0"
