+++ paper/paper.md
@@ -19,13 +19,13 @@ bibliography: paper.bib
 
 # Summary
 
-Solvent accessible surface area (SASA) calculations are fundamental for understanding protein structure, function, and dynamics in computational biology. These calculations quantify the surface area of biomolecules accessible to solvent molecules, providing insights into protein folding, stability, and intermolecular interactions. The Shrake-Rupley algorithm has served as the standard for SASA calculations since 1973, but existing implementations often become computational bottlenecks when analyzing large protein datasets. As proteomics datasets continue to grow exponentially—with initiatives like AlphaFold producing hundreds of millions of predicted protein structures—the need for efficient SASA calculation tools has become increased dramatically. RustSASA addresses this challenge by providing a high-performance implementation of the Shrake-Rupley algorithm written in pure Rust, delivering a 7× speed improvement over Freesasa while maintaining calculation accuracy and providing interfaces for multiple programming languages.
+Solvent accessible surface area (SASA) calculations are fundamental for understanding protein structure, function, and dynamics in computational biology. These calculations quantify the surface area of biomolecules accessible to solvent molecules, providing insights into protein folding, stability, and intermolecular interactions. The Shrake-Rupley algorithm has served as the standard for SASA calculations since 1973, but existing implementations often become computational bottlenecks when analyzing large protein datasets. As proteomics datasets continue to grow exponentially—with initiatives like AlphaFold producing hundreds of millions of predicted protein structures—the need for efficient SASA calculation tools has become increased dramatically. RustSASA addresses this challenge by providing a high-performance implementation of the Shrake-Rupley algorithm written in pure Rust, delivering a 7× speed improvement over Freesasa while maintaining calculation accuracy and providing interfaces for multiple programming languages and frameworks (i.e: MDAnalysis).
 
 # Statement of need
 
 Current SASA calculation tools represent a significant computational bottleneck in structural biology workflows, particularly for molecular dynamics simulations and high-throughput analyses. Popular implementations such as those in Biopython and Freesasa, while accurate, become prohibitively slow when processing large protein datasets.
 
-RustSASA addresses this performance gap by leveraging Rust's zero-cost abstractions and memory safety guarantees. Benchmarking on representative protein structures demonstrates that RustSASA achieves a 7× improvement over Freesasa and a 46× improvement over Biopython implementations. This performance advantage reduces computational costs for high-throughput structural analyses and makes large-scale comparative studies feasible. Furthermore, RustSASA's multi-language support (Rust and Python) and command-line interface ensure broad accessibility across the computational biology community.
+RustSASA addresses this performance gap by leveraging Rust's zero-cost abstractions and memory safety guarantees. Benchmarking on representative protein structures demonstrates that RustSASA achieves a 7× improvement over Freesasa, and a 46× performance improvement over Biopython. This performance advantage reduces computational costs for high-throughput structural analyses and makes large-scale comparative studies feasible. Furthermore, RustSASA's multi-language support (Rust and Python), command-line interface, and MDAnalysis package ensure broad accessibility across the computational biology community.
 
 # Results
 
@@ -42,16 +42,16 @@ RustSASA produces SASA values that closely match those from Freesasa, achieving 
 
 ![Comparing Freesasa, RustSasa, and Biopython performance on E. coli proteome\label{fig:example}](eval/performance_comparison.pdf){ width=100% }
 
[…]
