+++ Cargo.toml
@@ -15,10 +15,10 @@ rust-version = "1.85.0"
 pdbtbx = { path = "pdbtbx", version = "0.12.0" }             # We have to use a git module because you cannot use a git+rev version in a published crate for some reason...
 nalgebra = "0.33.2"
 rayon = "1.11.0"
-snafu = "0.8.8"
-clap = { version = "4.5.46", features = ["derive"] }
-serde = "1.0.219"
-serde_json = "1.0.143"
+snafu = "0.8.9"
+clap = { version = "4.5.48", features = ["derive"] }
+serde = "1.0.228"
+serde_json = "1.0.145"
 quick-xml = { version = "0.38.3", features = ["serialize"] }
 indicatif = "0.18.0"
 mimalloc = { version = "0.1.48", default-features = false }
