+++ .github/workflows/rust.yml
@@ -98,6 +98,7 @@ jobs:
 
   build:
     name: Build PGO optimized binaries
+    if: startsWith(github.ref, 'refs/tags/v')
     needs: [quality, test]
     strategy:
       matrix:
