### 2025-09-30 02a6bed
remove eroenous comment
