+++ src/main.rs
@@ -42,7 +42,6 @@ impl OutputFormat {
     }
 }
 
-/// Simple program to greet a person
 #[derive(Parser, Debug)]
 #[command(version, about, long_about = None)]
 struct Args {
