### 2025-09-30 e746a5c
fix paper .bib
