+++ paper/paper.bib
@@ -75,7 +75,7 @@ url={https://doi.org/10.1038/s41586-021-03819-2}
   number={11},
   pages={1422--1423},
   year={2009},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
   doi={10.1093/bioinformatics/btp163}
 }
 
