### 2025-10-27 23099f3
Bump indicatif from 0.18.0 to 0.18.2
Bumps [indicatif](https://github.com/console-rs/indicatif) from 0.18.0 to 0.18.2.
- [Release notes](https://github.com/console-rs/indicatif/releases)
- [Commits](https://github.com/console-rs/indicatif/commits)

---
updated-dependencies:
- dependency-name: indicatif
  dependency-version: 0.18.2
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
