### 2025-11-06 b3bedaf
- Replace SIMD with pulp theoretically improving performance on x86 CPUs with AVX SIMD extensions - Better support for PDB files with bad records - Fixed spelling of Niccolò Bruciaferri's name in paper Acknowledgements - Fixed PGO binary build step
