+++ src/options.rs
@@ -407,7 +407,7 @@ impl<T: SASAProcessor> SASAOptions<T> {
     /// ```
     /// use pdbtbx::StrictnessLevel;
     /// use rust_sasa::options::{SASAOptions, ResidueLevel};
-    /// let (mut pdb, _errors) = pdbtbx::open("./example.cif").unwrap();
+    /// let (mut pdb, _errors) = pdbtbx::open("./pdbs/example.cif").unwrap();
     /// let result = SASAOptions::<ResidueLevel>::new().process(&pdb);
     /// ```
     pub fn process(&self, pdb: &PDB) -> Result<T::Output, SASACalcError> {
