+++ Cargo.toml
@@ -22,6 +22,7 @@ serde_json = "1.0.145"
 quick-xml = { version = "0.38.3", features = ["serialize"] }
 indicatif = "0.18.0"
 mimalloc = { version = "0.1.48", default-features = false }
+pulp = "0.21.5"
 
 [dev-dependencies]
 approx = "0.5.1"
