### 2025-11-06 d001b41
Switch to dtolnay/rust-toolchain@stable and update actions/checkout@v4 to @v5
