+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rust-sasa"
-version = "0.3.4"
+version = "0.4.0"
 edition = "2024"
 description = "RustSASA is a Rust library for computing the absolute solvent accessible surface area (ASA/SASA) of each atom in a given protein structure using the Shrake-Rupley algorithm."
 readme = "README.md"
