### 2025-11-06 1c5672a
remove alphafold structures
