+++ paper/eval/download_e_coli_dataset.sh
@@ -1,0 +1,1 @@
+wget https://ftp.ebi.ac.uk/pub/databases/alphafold/latest/UP000000625_83333_ECOLI_v6.tar
