### 2025-11-07 ddd136b
nicer figures
