+++ paper/eval/perf_graphs.py
@@ -1,7 +1,3 @@
-rust_sasa_points = [8.071]
-freesasa_points = [54.914]
-biopython_points = [368.025]
-
 import matplotlib.pyplot as plt
 
 # Data for plotting
@@ -9,21 +5,56 @@ libraries = ["rust-sasa", "freesasa", "biopython"]
 times = [8.071, 54.914, 368.025]
 errors = [0.361, 0.455, 51.156]
 
-# Create the plot
-plt.figure(figsize=(10, 6))
-plt.rcParams.update({"font.size": 16})
-plt.bar(libraries, times, yerr=errors, color=["#33BBEE", "#EE7733", "#EE3377"], capsize=5)
-plt.ylabel("Time (seconds)")
-plt.title("Performance Comparison of SASA Libraries on E. coli proteome")
-# plt.yscale("log")  # Using log scale due to large differences in values
+# --- Styling parameters ---
+fontsize = 18
+linewidth = 2.5
+bar_colors = ["#33BBEE", "#EE7733", "#EE3377"]  # Paul Tol color-blind-safe palette
 
-# Add value labels on bars
-for i, v in enumerate(times):
-    plt.text(i, v + errors[i], f"{v}s", ha="center", va="bottom")
+# --- Create figure and axes ---
+fig, ax = plt.subplots(figsize=(10, 6))
 
-# Add buffer space above bars to prevent text from hitting figure top
+# --- Axes styling ---
+ax.spines["right"].set_visible(False)
+ax.spines["top"].set_visible(False)
+ax.spines["left"].set_linewidth(linewidth)
+ax.spines["bottom"].set_linewidth(linewidth)
+ax.tick_params(axis="both", width=linewidth, length=8, direction="out")
+
+# --- Bar chart (no outlines) ---
+bars = ax.bar(
+    libraries,
+    times,
+    yerr=errors,
+    color=bar_colors,
+    capsize=6,
+)
+
+# --- Labels ---
+ax.set_ylabel("Time (seconds)", fontsize=fontsize + 2)
+ax.set_title("Performance Comparison of SASA Libraries on E. coli Proteome", fontsize=fontsize + 4, pad=15)
+ax.tick_params(axis="x", labelsize=fontsize)
+ax.tick_params(axis="y", labelsize=fontsize)
+
+# --- Optional: log scale (uncomment if large dynamic range) ---
+# ax.set_yscale("log")
+
+# --- Value labels above bars ---
+for bar, err, val in zip(bars, errors, times):
+    height = bar.get_height()
+    ax.text(
+        bar.get_x() + bar.get_width() / 2,
+        height + err + (0.02 * max(times)),
+        f"{val:.1f}s",
+        ha="center",
+        va="bottom",
+        fontsize=fontsize - 2,
+    )
+
+# --- Grid and limits ---
 max_height = max(times[i] + errors[i] for i in range(len(times)))
-plt.ylim(0, max_height * 1.10)
+ax.set_ylim(0, max_height * 1.15)
+ax.grid(True, axis="y", linestyle="--", alpha=0.3)
 
+# --- Layout and save ---
 plt.tight_layout()
-plt.savefig("performance_comparison.pdf")
+plt.savefig("figures/performance_comparison.pdf", bbox_inches="tight", dpi=300)
