+++ paper/eval/freesasa_dataset.txt
@@ -1,88 +1,0 @@
-1d8w
-1fo8
-1g5a
-1gqi
-1h16
-1hbn
-1jcd
-1jz8
-1kqf
-1m15
-1mj5
-1n62
-1o6g
-1oew
-1q6z
-1rk6
-1rwh
-1su8
-1t1u
-2bz1
-2cvi
-2d5w
-2drt
-2dvm
-2e4t
-2e7z
-2eab
-2gb4
-2gj4
-2gpi
-2heu
-2hhc
-2i53
-2odk
-2olr
-2qud
-2yfo
-2zux
-3a5f
-3b0x
-3bon
-3bvx
-3cuz
-3fo3
-3fss
-3g02
-3hkw
-3i94
-3ie7
-3kyz
-3lqb
-3m0z
-3og2
-3pzw
-3sqz
-3t47
-3tew
-3tg7
-3u9w
-3uc7
-3w7t
-3w7y
-3wa2
-3weo
-3whr
-3znv
-3zsj
-4a9v
-4bb9
-4bps
-4btv
-4c1a
-4cj0
-4g6t
-4g7x
-4gmu
-4h7u
-4kv7
-4la2
-4lix
-4n13
-4oh7
-4oxx
-4p0t
-4pj2
-4qas
-4xfj
-7odc
