### 2025-11-07 d48b8cd
Nicer figures
