### 2025-11-17 355a470
Bump indicatif from 0.18.2 to 0.18.3
Bumps [indicatif](https://github.com/console-rs/indicatif) from 0.18.2 to 0.18.3.
- [Release notes](https://github.com/console-rs/indicatif/releases)
- [Commits](https://github.com/console-rs/indicatif/compare/0.18.2...0.18.3)

---
updated-dependencies:
- dependency-name: indicatif
  dependency-version: 0.18.3
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
