### 2025-11-17 03ef838
Bump quick-xml from 0.38.3 to 0.38.4
Bumps [quick-xml](https://github.com/tafia/quick-xml) from 0.38.3 to 0.38.4.
- [Release notes](https://github.com/tafia/quick-xml/releases)
- [Changelog](https://github.com/tafia/quick-xml/blob/master/Changelog.md)
- [Commits](https://github.com/tafia/quick-xml/compare/v0.38.3...v0.38.4)

---
updated-dependencies:
- dependency-name: quick-xml
  dependency-version: 0.38.4
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
