### 2025-11-25 maxall41/RustSASA#40 @mittinatten
Benchmarks are somewhat misleading
I am reporting as part of the review process for https://github.com/openjournals/joss-reviews/issues/8727.

First I want to congratulate you on a nice piece of software, I think for a lot of users this could be a relevant replacement for FreeSASA. I am the author of FreeSASA so I will focus on the benchmarks against FreeSASA here and then probably submit some separate issues in the coming days.

I was very impressed by the numbers you presented when comparing with FreeSASA. However, although I am not a Rust programmer, I couldn't see anything in the code that would explain the difference. It looks like the actual calculation is almost identical to the C code, except for a smarter parallelization of the atom-by-atom calculation, so I was surprised by it being 7 times faster. I would have thought that to have any significant effect on performance either a completely new algorithm is needed, or some fancy SIMD instructions.

## Calculation parameters

I therefore had a look at the benchmark code: As far as I can tell from the test-script for FreeSASA you are running it with the default parameters, which have higher precision than the defaults in RustSASA, and therefore lead to slower calculations. So the comparison is a little bit apples to oranges.

FreeSASA can be parameterized to have the same precision as the RustSASA defaults using
```sh
freesasa --shrake-rupley -n 100 --format json --output [output.json] [input.pdb]
```

When I run the comparison for a single file using those settings, the performance is about the same (FreeSASA seems to be about 5-10% faster).

As a side note, as far as I can remember from when I wrote FreeSASA, you get a little more precision per spent CPU cycle using the Lee & Richards algorithm, but it is also quite a lot more cumbersome to implement. But they are parameterized differently, so it's hard to say exactly which number of Shrake and Rupley points correspond to which number of Lee and Richards slices, in terms of precision.

## Parallelization

When running over a whole directory at once, RustSASA is still much faster than FreeSASA parallelized using `parallel`, also after using the same parameters. I assume this is because `parallel` has an overhead for creating a process for each calculation, and maybe isn't as efficient at utilizing all cores. (In my runs with `parallel`, FreeSASA is more than 10 times slower than RustSASA in folder mode, so maybe I'm doing something wrong).

To see how much could be blamed on `parallel` (or my use thereof), I wrote a simple Node.js script using [p-limit](https://www.npmjs.com/package/p-limit) to parallelize the runs instead (see below). In that case RustSASA in folder mode is about 2.5 times faster than FreeSASA.

In my script I run 1. FreeSASA one file at a time, 2. RustSASA one file at a time, and 3. RustSASA for the whole folder at once. Here is the output from a sample run on the E. Coli dataset.

```
Freesasa parallel processes  8.78  seconds
Rustsasa parallel processes  9.90  seconds
Rustsasa single batch process  3.83  seconds
```

The results are about the same for every run on my machine (an M3 Mac), with a variance of 5% maybe. As you can see when RustSASA is run one file at a time, it performs more or less at the same level as FreeSASA.

## Requested changes to paper

So, in conclusion: I would like the paper to reflect these nuances and the graphs there and in the Readme to be updated accordingly. For some workflows the ability to run very fast calculations over a whole folder of files will be an important improvement over using FreeSASA, but for other pipelines the single file performance is what matters (and then I guess it depends on how well the two technologies fit into your stack as a whole, or if you need the larger feature set in FreeSASA). I think the paper at the moment doesn't explain why and in which scenarios RustSASA is faster. Rust as a language shouldn't have any performance benefit over C on a general basis.

## My test code

Feel free to use my script or a variation thereof for your analysis, if you find it useful. I didn't spend many minutes on it though, it's just meant as a PoC, and I'm sure there are other approaches and runtimes that are more efficient than Node.js.

```ts
// run this with Node 22.19 or higher, in a directory with p-limit installed, store as a .ts-file.
// command:  node index.ts <input-directory> <output-directory>

import pLimit from 'p-limit';
import fs from 'node:fs';
import path from 'node:path';
import { exec as execWithCb } from 'node:child_process';
import { promisify } from 'node:util';
const exec = promisify(execWithCb);

const input = process.argv[2];
const output = process.argv[3];

if (!input || !output) {
  console.log("usage: node index.ts <input-dir> <output-dir>");
  process.exit();
}

const files = fs.readdirSync(input);

const limit = pLimit(8);

const freesasaTasks = files.map(file => limit(() => {
  const inputFile = path.join(input, file);
  const outputFile = path.join(output, file.replace(".pdb", ".freesasa.json"));
  return exec(`freesasa --shrake-rupley -n 100 -t 2 --output ${outputFile} --format json ${inputFile}`);
}));

const rustsasaTasks = files.map(file => limit(() => {
  const inputFile = path.join(input, file);
  const outputFile = path.join(output, file.replace(".pdb", ".rustsasa.json"));
  return exec(`~/.cargo/bin/rust-sasa ${inputFile} ${outputFile} --format=json`);
}));

const startFreesasa = performance.now();
await Promise.all(freesasaTasks);
console.log("Freesasa parallel processes ", ((performance.now() - startFreesasa) / 1000).toFixed(2), " seconds");

const startRust = performance.now();
await Promise.all(rustsasaTasks);
console.log("Rustsasa parallel processes ", ((performance.now() - startRust) / 1000).toFixed(2), " seconds");

const startRust2 = performance.now();
await exec(`~/.cargo/bin/rust-sasa ${input} ${output} --format=json`);
console.log("Rustsasa single batch process ", ((performance.now() - startRust2) / 1000).toFixed(2), " seconds");
```
