### 2025-11-26 bf02ab1
Faster spatial grid implementation
