### 2025-11-26 5d0e914
Faster spatial grid implementation
