+++ src/structures/spatial_grid.rs
@@ -1,17 +1,27 @@
 use crate::Atom;
 
-/// Fast 3D grid for spatial queries
+/// Spatial grid for optimized search
 #[repr(C)]
 pub(crate) struct SpatialGrid {
-    grid: Vec<Vec<Vec<Vec<usize>>>>, // 3D grid of atom indices
-    min_bounds: [f32; 3],            // 12 bytes
-    cell_size: f32,                  // 4 bytes
-    grid_size: [usize; 3],           // 24 bytes
+    /// A single contiguous buffer of atom indices, sorted by grid cell.
+    sorted_atom_indices: Vec<usize>,
+    /// A parallel buffer of positions (x, y, z) corresponding to `sorted_atom_indices`.
+    sorted_positions: Vec<f32>,
+    /// The start index in `sorted_atom_indices` for each flat cell index.
+    cell_starts: Vec<u32>,
+    /// min bounds for grid
+    min_bounds: [f32; 3],
+    /// Size of each cell
+    cell_size: f32,
+    /// Dimensions of grid
+    grid_dims: [usize; 3],
+    /// Dimension strides for flattened atom vec
+    strides: [usize; 3],
 }
 
 impl SpatialGrid {
     pub(crate) fn new(atoms: &[Atom], cell_size: f32) -> Self {
-        // Calculate bounds
+        // Calculate Bounds
         let mut min_bounds = [f32::INFINITY; 3];
         let mut max_bounds = [f32::NEG_INFINITY; 3];
 
@@ -23,85 +33,153 @@ impl SpatialGrid {
             }
         }
 
-        // Add padding
+        // Add padding to bounds
         for i in 0..3 {
             min_bounds[i] -= cell_size;
             max_bounds[i] += cell_size;
         }
 
         // Calculate grid size
-        let grid_size = [
+        let grid_dims = [
             ((max_bounds[0] - min_bounds[0]) / cell_size).ceil() as usize + 1,
             ((max_bounds[1] - min_bounds[1]) / cell_size).ceil() as usize + 1,
             ((max_bounds[2] - min_bounds[2]) / cell_size).ceil() as usize + 1,
         ];
 
-        // Initialize grid
-        let mut grid = vec![vec![vec![Vec::new(); grid_size[2]]; grid_size[1]]; grid_size[0]];
+        // Use stride to index into 3D grid as 1D vec
+        let strides = [1, grid_dims[0], grid_dims[0] * grid_dims[1]];
+        let num_cells = grid_dims[0] * grid_dims[1] * grid_dims[2];
 
-        // Add atoms to grid
+        // Count number of atoms in each cell for future indexing into cells
+        let mut cell_counts = vec![0u32; num_cells];
+        let inv_cell_size = 1.0 / cell_size;
+
+        // Helper to get flat index using strides
+        let get_cell_idx = |pos: &[f32; 3]| -> usize {
+            let x = ((pos[0] - min_bounds[0]) * inv_cell_size) as usize;
+            let y = ((pos[1] - min_bounds[1]) * inv_cell_size) as usize;
+            let z = ((pos[2] - min_bounds[2]) * inv_cell_size) as usize;
+            x * strides[0] + y * strides[1] + z * strides[2]
+        };
+
+        for atom in atoms {
+            let pos = [atom.position.x, atom.position.y, atom.position.z];
+            let idx = get_cell_idx(&pos);
+            cell_counts[idx] += 1
+        }
+
+        // Create cell start index using cell_counts
+        let mut cell_starts = vec![0u32; num_cells + 1];
+        let mut current_offset = 0;
+        for i in 0..num_cells {
+            cell_starts[i] = current_offset;
+            current_offset += cell_counts[i];
[…]
