+++ paper/eval/benchmark.sh
@@ -1,0 +1,2 @@
+ hyperfine --warmup 3 --runs 3 "sh exec_rustsasa.sh ./data/UP000000625_83333_ECOLI_v6/ ./results/E_coli_RSASA/"
+ hyperfine --warmup 3 --runs 3 "sh exec_freesasa.sh ./data/UP000000625_83333_ECOLI_v6/ ./results/E_coli_RSASA/"
