### 2025-11-26 30da271
neighbor sorting optimization
