### 2025-11-27 13c8bef
custom simd
