+++ src/lib.rs
@@ -20,14 +20,28 @@ pub use crate::structures::atomic::*;
 
 use structures::spatial_grid::SpatialGrid;
 // Re-export io functions for use in the binary crate
+use pulp::Arch;
+use rayon::prelude::*;
 pub use utils::io::{sasa_result_to_json, sasa_result_to_protein_object, sasa_result_to_xml};
 
-use nalgebra::{Point3, Vector3};
-use rayon::prelude::*;
+struct SpherePointsSoA {
+    x: Vec<f32>,
+    y: Vec<f32>,
+    z: Vec<f32>,
+}
+
+impl SpherePointsSoA {
+    fn len(&self) -> usize {
+        self.x.len()
+    }
+}
 
 /// Generates points on a sphere using the Golden Section Spiral algorithm
-fn generate_sphere_points(n_points: usize) -> Vec<Vector3<f32>> {
-    let mut points = Vec::with_capacity(n_points);
+fn generate_sphere_points(n_points: usize) -> SpherePointsSoA {
+    let mut x = Vec::with_capacity(n_points);
+    let mut y = Vec::with_capacity(n_points);
+    let mut z = Vec::with_capacity(n_points);
+
     let inv_n_points = 1.0 / n_points as f32;
 
     for i in 0..n_points {
@@ -40,38 +54,12 @@ fn generate_sphere_points(n_points: usize) -> Vec<Vector3<f32>> {
         let (sin_azimuth, cos_azimuth) = azimuth.sin_cos();
         let sin_inclination = inclination.sin();
 
-        let x = sin_inclination * cos_azimuth;
-        let y = sin_inclination * sin_azimuth;
-        let z = inclination.cos();
-
-        points.push(Vector3::new(x, y, z));
+        x.push(sin_inclination * cos_azimuth);
+        y.push(sin_inclination * sin_azimuth);
+        z.push(inclination.cos());
     }
 
-    points
-}
-
-fn is_accessible_precomputed(
-    test_point: &Point3<f32>,
-    atom: &Atom,
-    neighbors: &[NeighborData],
-    atoms: &[Atom],
-) -> bool {
-    for neighbor_data in neighbors {
-        let neighbor = &atoms[neighbor_data.idx as usize];
-        if atom.id != neighbor.id {
-            // Manual distance calculation is faster than nalgebra's norm_squared
-            let pos = neighbor.position.coords.xyz();
-            let test_pos = test_point.coords.xyz();
-            let dx = test_pos[0] - pos[0];
-            let dy = test_pos[1] - pos[1];
-            let dz = test_pos[2] - pos[2];
-            let dist_sq = dx * dx + dy * dy + dz * dz;
-            if dist_sq < neighbor_data.threshold_squared {
-                return false;
-            }
-        }
-    }
-    true
+    SpherePointsSoA { x, y, z }
 }
 
 #[inline(never)]
@@ -131,6 +119,120 @@ fn precompute_neighbors(
     neighbors
 }
 
+struct AtomSasaKernel<'a> {
+    atom_index: usize,
+    atoms: &'a [Atom],
+    neighbors: &'a [NeighborData],
+    sphere_points: &'a SpherePointsSoA,
+    probe_radius: f32,
+}
+
+impl<'a> pulp::WithSimd for AtomSasaKernel<'a> {
+    type Output = f32;
+
+    #[inline(always)]
+    fn with_simd<S: pulp::Simd>(self, simd: S) -> Self::Output {
+        let atom = &self.atoms[self.atom_index];
+        let center_pos = atom.position;
+        let r = atom.radius + self.probe_radius;
+        let r2 = r * r;
+
[…]
