### 2025-11-27 dbaef6e
fix clippy issue
