### 2025-11-27 1744641
- Remove support opt - Save Hyperfine results to JSON and load to create perf graphs - Change exec_freesasa params to match Rustsasa params - Support --n-points customization when using CLI (default: 100)
