### 2025-11-27 68bb331
Update benchmarks, paper, and readme.
