+++ paper/eval/perf_graphs.py
@@ -3,91 +3,117 @@ import os
 
 import matplotlib.pyplot as plt
 
-# --- Load data from JSON files ---
-json_files = [
+multi_data_fiiles = [
     "results/bench_rustsasa.json",
     "results/bench_freesasa.json",
     "results/bench_biopython.json",
 ]
+single_data_fiiles = [
+    "results/bench_rustsasa_single.json",
+    "results/bench_freesasa_single.json",
+    "results/bench_biopython_single.json",
+]
 
-data_entries = []
 
-for json_file in json_files:
-    if not os.path.exists(json_file):
-        print(f"Warning: {json_file} not found. Skipping.")
-        continue
+def create_graph(json_files, output_file, title):
+    data_entries = []
+    for json_file in json_files:
+        if not os.path.exists(json_file):
+            print(f"Warning: {json_file} not found. Skipping.")
+            continue
+        with open(json_file) as f:
+            data = json.load(f)
+            result = data["results"][0]
+            command = result["command"]
+            mean_time = result["mean"]
+            stddev = result["stddev"]
+            label = "Unknown"
+            if "rustsasa" in command or "rust-sasa" in command:
+                label = "rust-sasa"
+            elif "freesasa" in command:
+                label = "freesasa"
+            elif "biopython" in command:
+                label = "biopython"
+            data_entries.append(
+                {
+                    "label": label,
+                    "time": mean_time,
+                    "error": stddev,
+                },
+            )
 
-    with open(json_file) as f:
-        data = json.load(f)
-        result = data["results"][0]
+    # Sort order: rust-sasa, freesasa, biopython
+    order = {"rust-sasa": 0, "freesasa": 1, "biopython": 2}
+    data_entries.sort(key=lambda x: order.get(x["label"], 999))
 
-        command = result["command"]
-        mean_time = result["mean"]
-        stddev = result["stddev"]
+    libraries = [d["label"] for d in data_entries]
+    times = [d["time"] for d in data_entries]
+    errors = [d["error"] for d in data_entries]
 
-        label = "Unknown"
-        if "rustsasa" in command:
-            label = "rust-sasa"
-        elif "freesasa" in command:
-            label = "freesasa"
-        elif "biopython" in command:
-            label = "biopython"
+    # Determine if we should use milliseconds
+    use_milliseconds = any(t < 1.0 for t in times)
 
-        data_entries.append(
-            {
-                "label": label,
-                "time": mean_time,
-                "error": stddev,
-            },
-        )
+    if use_milliseconds:
+        times = [t * 1000 for t in times]
+        errors = [e * 1000 for e in errors]
+        time_unit = "milliseconds"
+        time_suffix = "ms"
+        decimal_places = 0
+    else:
+        time_unit = "seconds"
+        time_suffix = "s"
+        decimal_places = 1
 
-# Sort order: rust-sasa, freesasa, biopython
-order = {"rust-sasa": 0, "freesasa": 1, "biopython": 2}
-data_entries.sort(key=lambda x: order.get(x["label"], 999))
+    fontsize = 18
+    linewidth = 2.5
+    bar_colors = ["#33BBEE", "#EE7733", "#EE3377"]  # Paul Tol color-blind-safe palette
 
-libraries = [d["label"] for d in data_entries]
-times = [d["time"] for d in data_entries]
[…]

[…]
