### 2025-11-27 91a514a
Update benchmarks, paper, and readme.
