### 2025-11-27 maxall41/RustSASA#40 @maxall41
Thanks for the great feedback! I have corrected the benchmarking code to resolve this issue, and I have updated the paper and readme to reflect the new results. Additionally, I have made the following performance improvements in v0.5.0 of RustSASA which deliver a ~30% improvement in performance over v0.4.0:
* Smarter spatial grid data structure.
* Custom SIMD Kernel to take full advantage of modern instructions.
* Sorting precomputed neighbor atoms by distance to search atom to increase probability of early exit in `is_accessible_precomputed`.

With all of these improvements, and the updated benchmarking code, the performance improvement over Freesasa is ~5x when calculating SASA values for the entire AlphaFold E. coli proteome. And RustSASA's performance is now roughly equivalent to that of Freesasa on single proteins.
