### 2025-11-27 1a84cfe
Update paper and README to reflect increase in performance over Biopython with v0.5.0. And update readme to reflect PGO optimized build performance.
