### 2025-11-27 #8727 @eunos-1128
@prashjha

Thank you for considering me as a potential reviewer. My primary expertise lies in protein structure prediction and small-molecule ligand docking. Although I have experience in developing and packaging tools like FreeSASA via Homebrew and Conda, my background does not encompass Rust programming or high-performance computational implementation at a deep technical level.

Given this, I am uncertain whether I possess the necessary expertise to evaluate the technical accuracy and performance aspects of the RustSASA implementation in this manuscript. I want to ensure a thorough and fair review, so I would appreciate your guidance on whether my expertise aligns with the review requirements for this work.
