### 2025-11-27 7a071c9
accelerated spatial grid search
