### 2025-11-28 c4cd31a
tighter bounds
