+++ src/utils.rs
@@ -4,9 +4,9 @@ use std::sync::LazyLock;
 
 use pulp::Arch;
 
-static ARCH: LazyLock<Arch> = LazyLock::new(Arch::new);
+pub(crate) static ARCH: LazyLock<Arch> = LazyLock::new(Arch::new);
 
-pub fn simd_sum(values: &[f32]) -> f32 {
+pub(crate) fn simd_sum(values: &[f32]) -> f32 {
     let mut total = 0f32;
     ARCH.dispatch(|| {
         for x in values {
