+++ src/lib.rs
@@ -17,10 +17,10 @@ mod utils;
 
 pub use crate::options::*;
 pub use crate::structures::atomic::*;
+use crate::utils::ARCH;
 
 use structures::spatial_grid::SpatialGrid;
 // Re-export io functions for use in the binary crate
-use pulp::Arch;
 use rayon::prelude::*;
 pub use utils::io::{sasa_result_to_json, sasa_result_to_protein_object, sasa_result_to_xml};
 
@@ -273,11 +273,9 @@ pub fn calculate_sasa_internal(
     // Use precomputed neighbors
     let neighbor_indices = precompute_neighbors(atoms, probe_radius, max_radii);
 
-    let arch = Arch::new();
-
     // Helper closure to wrap the kernel dispatch
     let process_atom = |(i, neighbors): (usize, &Vec<NeighborData>)| {
-        arch.dispatch(AtomSasaKernel {
+        ARCH.dispatch(AtomSasaKernel {
             atom_index: i,
             atoms,
             neighbors,
