### 2025-12-03 #8727 @prashjha
Hi @kliment-olechnovic, I am looking for a second reviewer for the JOSS submission above, and this work may be in your domain, as suggested by @mlund. Would you be interested in reviewing this? You can find the PDF at this link [pdf link](https://github.com/openjournals/joss-papers/blob/joss.08727/joss.08727/10.21105.joss.08727.pdf), and the corresponding software is at this [link](https://github.com/maxall41/RustSASA).

This [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) covers almost everything about the review process; I will be happy to help with more information.

Please let me know if you are interested. Thank you!
