### 2025-12-03 #8727 @kliment-olechnovic
Hi @prashjha , I can review this submission, if being proficient in Rust is not a strict requirement (I have never done anything in Rust).
