### 2025-12-03 #9537 @editorialbot
[REVIEW]: RustSASA: A Rust Crate for Accelerated Solvent Accessible Surface Area Calculations
**Submitting author:** @maxall41 (<a href="http://orcid.org/0000-0002-0959-1164">Maxwell James Campbell</a>)
**Repository:** https://github.com/maxall41/RustSASA
**Branch with paper.md** (empty if default branch):
**Version:** v0.9.0
