### 2025-12-04 maxall41/RustSASA#44 @mittinatten
How are hydrogens and atomic radii handled?
As far as I can tell from reading the code, there is no special handling of hydrogen atoms.  A lot of experimental protein structures don't have hydrogen atoms, and to compensate for this, it is normal to use atomic radii that are larger than the VdW radius of the corresponding elements in the SASA calculation. The radius will then vary for the same element depending on the chemical environment. If you have a structure with hydrogens, you would then either have to ignore the hydrogens (which makes the calculation a lot faster) or use another set of atomic radii. If you have a set of radii that are optimized for structures _with_ hydrogens, you could inject likely hydrogen positions into the structure when they are missing (they should be pretty straightforward to approximate), but that would make the calculation slower instead.

There is probably no one good answer to how to approach this, SASA is after all an approximate concept with no "true" value, to begin with. The important thing is to be consistent, so you can compare between calculations. If you are interested in per atom SASA, both stripping and adding hydrogens will have consequences when mapping back to the original structure.

The solution in FreeSASA, for comparison, is that there are three different consistent sets of atomic radii that you can choose from. They have all been calculated to work with hydrogen free structures, and the FreeSASA documentation states that if you include hydrogens (which can be done by a command line flag), you should provide your own set of atomic radii that are consistent with the hydrogens being included. The defaults are the atomic radii by [Tsai et al.](https://pubmed.ncbi.nlm.nih.gov/10388571/), called [_ProtOr_](https://github.com/mittinatten/freesasa/blob/master/share/protor.config).

If you never have hydrogens in your structures, or always have them, the error will be always be in the same direction, and might not matter, depending on your application (this would be the case for computed structures, from simulations for example). However, if you mix both types of structures (i.e. have experimental structures), it will be difficult to compare between structures or calculate relative SASA, etc.

Perhaps this is already handled by the third-party PDB-reader, or I overlooked something in the code, but I would at least like a clarification of your thoughts here. If there is no clear strategy already, I think one is necessary. The good news is that if you choose to strip hydrogens (and haven't already) you could see a quite significant reduction in computation time for some structure.

Linked to the JOSS review issue: https://github.com/openjournals/joss-reviews/issues/9537
