### 2025-12-05 bc4411b
add instructions for building from source
