### 2025-12-05 35b387d
Fix comment & updated paper verbage
