### 2025-12-05 maxall41/RustSASA#40 @maxall41
Thanks! Firstly, the `is_accessible_precomputed` function was rolled into the SIMD kernel implementation. I have updated the comment to fix the erroneous mention of the function. Secondly, I have updated the first paragraph of the "Statement of need" section to read as follows:

> Current SASA calculation tools represent a significant computational bottleneck in structural biology workflows, particularly for molecular dynamics simulations and high-throughput analyses. Popular implementations such as those in Biopython and Freesasa, while accurate, become prohibitively slow when processing large protein datasets. RustSASA addresses this performance gap by leveraging Rust's efficient parallelization abstractions (via Rayon) and readily available SIMD instructions (via Pulp). These optimizations enable RustSASA's performance advantage over the simpler C implementation of the same algorithm in Freesasa.
