+++ src/test.rs
@@ -478,6 +478,8 @@ const FIXED_HIGH_RES_ATOMS: [f32; 2622] = [
 mod tests {
     use super::*;
     use crate::options::SASAOptions;
+    use crate::structures::spatial_grid::SpatialGrid;
+    use crate::utils::consts::{PROTOR_RADII, get_protor_radius};
     use crate::{Atom, AtomLevel, ChainLevel, ProteinLevel, ResidueLevel, calculate_sasa_internal};
     use approx::assert_abs_diff_eq;
     use nalgebra::Point3;
@@ -489,24 +491,21 @@ mod tests {
         let (pdb, _errors) = pdbtbx::open("./pdbs/example.cif").unwrap();
         let mut atoms = vec![];
         for atom in pdb.atoms() {
+            let element = atom.element().unwrap();
             atoms.push(Atom {
                 position: Point3::new(
                     atom.pos().0 as f32,
                     atom.pos().1 as f32,
                     atom.pos().2 as f32,
                 ),
-                radius: atom
-                    .element()
-                    .unwrap()
-                    .atomic_radius()
-                    .van_der_waals
-                    .unwrap() as f32,
+                radius: element.atomic_radius().van_der_waals.unwrap() as f32,
                 id: atom.serial_number(),
                 parent_id: None,
+                is_hydrogen: element == &pdbtbx::Element::H,
             })
         }
         let start = Instant::now();
-        let sasa = calculate_sasa_internal(&atoms, 1.4, 100, true);
+        let sasa = calculate_sasa_internal(&atoms, 1.4, 100, true, false);
         let duration = start.elapsed();
         // Compare element by element since Vec<f32> doesn't implement AbsDiffEq
         assert_eq!(sasa.len(), FIXED_LOW_RES_ATOMS.len());
@@ -588,11 +587,11 @@ mod tests {
         let duration = start.elapsed();
         println!("Time elapsed (ATOM): {duration:?}");
 
-        assert_abs_diff_eq!(protein_sasa.global_total, 20279.314, epsilon = 1500.0);
+        assert_abs_diff_eq!(protein_sasa.global_total, 21790.121, epsilon = 1500.0);
         assert_abs_diff_eq!(protein_sasa.polar_total, 4279.8906, epsilon = 1500.0);
         assert_abs_diff_eq!(protein_sasa.non_polar_total, 15999.43, epsilon = 1500.0);
 
-        assert_abs_diff_eq!(chain_sasa[0].value, 20279.314, epsilon = 1500.0);
+        assert_abs_diff_eq!(chain_sasa[0].value, 21790.121, epsilon = 1500.0);
         assert_eq!(chain_sasa[0].name, "A");
 
         // Compare element by element since Vec<f32> doesn't implement AbsDiffEq
@@ -604,8 +603,6 @@ mod tests {
 
     #[test]
     fn spatial_grid() {
-        use crate::structures::spatial_grid::SpatialGrid;
-
         // Create test atoms in a simple 2x2x2 grid pattern
         let atoms = vec![
             Atom {
@@ -613,29 +610,34 @@ mod tests {
                 radius: 1.0,
                 id: 1,
                 parent_id: None,
+                is_hydrogen: false,
             },
             Atom {
                 position: Point3::new(5.0, 0.0, 0.0),
                 radius: 1.0,
                 id: 2,
                 parent_id: None,
+                is_hydrogen: false,
             },
             Atom {
                 position: Point3::new(0.0, 5.0, 0.0),
                 radius: 1.0,
                 id: 3,
                 parent_id: None,
+                is_hydrogen: false,
             },
             Atom {
                 position: Point3::new(10.0, 10.0, 10.0),
                 radius: 1.0,
[…]

[…]
