+++ src/main.rs
@@ -64,6 +64,18 @@ struct Args {
     /// Number of Shrake Rupley points
     #[arg(short, long, default_value_t = 100)]
     n_points: usize,
+
+    /// Probe radius in Angstroms (default: 1.4)
+    #[arg(short, long, default_value_t = 1.4)]
+    probe_radius: f32,
+
+    /// Include hydrogen atoms in SASA calculation (default: hydrogens are excluded)
+    #[arg(short = 'H', long, default_value_t = false)]
+    include_hydrogens: bool,
+
+    /// Path to custom radii configuration file (default: uses embedded protor.config)
+    #[arg(short = 'r', long)]
+    radii_file: Option<String>,
 }
 
 #[derive(Debug, Snafu)]
@@ -94,6 +106,9 @@ pub enum CLIError {
 
     #[snafu(display("Failed to write output file"))]
     FileWrite { source: std::io::Error },
+
+    #[snafu(display("Failed to load radii file"))]
+    RadiiFileLoad { source: std::io::Error },
 }
 
 fn process(
@@ -102,14 +117,25 @@ fn process(
     output_depth: SASALevel,
     format: &OutputFormat,
     n_points: usize,
+    probe_radius: f32,
     parallel: bool,
+    include_hydrogens: bool,
+    radii_file: Option<&str>,
 ) -> Result<(), CLIError> {
     let (pdb, _) = ReadOptions::default()
         .set_level(pdbtbx::StrictnessLevel::Loose)
         .read(input_file)
         .map_err(|errors| CLIError::InputFileRead { errors })?;
-    let result = calculate_sasa_and_wrap(&pdb, &output_depth, n_points, parallel)
-        .context(SASACalculationSnafu)?;
+    let result = calculate_sasa_and_wrap(
+        &pdb,
+        &output_depth,
+        n_points,
+        probe_radius,
+        parallel,
+        include_hydrogens,
+        radii_file,
+    )
+    .context(SASACalculationSnafu)?;
 
     match format {
         OutputFormat::Xml => {
@@ -138,55 +164,95 @@ fn process(
     Ok(())
 }
 
-/// Backward compatibility helper function
+/// Macro to reduce duplication in calculate_sasa_and_wrap
+macro_rules! process_level {
+    ($level_constructor:ident, $result_variant:ident, $pdb:expr, $n_points:expr, $probe_radius:expr, $parallel:expr, $include_hydrogens:expr, $radii_file:expr) => {{
+        let mut options = SASAOptions::$level_constructor()
+            .with_parallel($parallel)
+            .with_n_points($n_points)
+            .with_probe_radius($probe_radius)
+            .with_include_hydrogens($include_hydrogens);
+        if let Some(path) = $radii_file {
+            options = options
+                .with_radii_file(path)
+                .map_err(|e| SASACalcError::RadiiFileLoad { source: e })?;
+        }
+        let result = options.process($pdb)?;
+        Ok(SASAResult::$result_variant(result))
+    }};
+}
+
 fn calculate_sasa_and_wrap(
     pdb: &pdbtbx::PDB,
     level: &SASALevel,
     n_points: usize,
+    probe_radius: f32,
     parallel: bool,
+    include_hydrogens: bool,
+    radii_file: Option<&str>,
 ) -> Result<SASAResult, SASACalcError> {
     match level {
-        SASALevel::Atom => {
-            let result = SASAOptions::atom_level()
-                .with_parallel(parallel)
-                .with_n_points(n_points)
-                .process(pdb)?;
-            Ok(SASAResult::Atom(result))
-        }
-        SASALevel::Residue => {
-            let result = SASAOptions::residue_level()
-                .with_parallel(parallel)
-                .with_n_points(n_points)
-                .process(pdb)?;
[…]
