### 2025-12-11 maxall41/RustSASA#44 @maxall41
I have updated RustSASA to handle hydrogens properly (as of v0.6.0). RustSASA now excludes hydrogens by default and uses the ProtOr radii. If users want to include hydrogens, they can do so by passing `--include-hydrogens true` into the CLI (or API equivalent) and specifying a custom radii file with `--radii-file`.
