+++ paper/paper.bib
@@ -92,3 +92,18 @@ url={https://doi.org/10.1038/s41586-021-03819-2}
   year = {2011},
   pages = {42-47}
 }
+
+@article{ProtOr,
+title = {The packing density in proteins: standard radii and volumes11Edited by J. M. Thornton},
+journal = {Journal of Molecular Biology},
+volume = {290},
+number = {1},
+pages = {253-266},
+year = {1999},
+issn = {0022-2836},
+doi = {https://doi.org/10.1006/jmbi.1999.2829},
+url = {https://www.sciencedirect.com/science/article/pii/S0022283699928292},
+author = {Jerry Tsai and Robin Taylor and Cyrus Chothia and Mark Gerstein},
+keywords = {atomic group radii and volumes, Voronoi polyhedra, packing efficiency, protein interior},
+abstract = {The sizes of atomic groups are a fundamental aspect of protein structure. They are usually expressed in terms of standard sets of radii for atomic groups and of volumes for both these groups and whole residues. Atomic groups, which subsume a heavy-atom and its covalently attached hydrogen atoms into one moiety, are used because the positions of hydrogen atoms in protein structures are generally not known. We have calculated new values for the radii of atomic groups and for the volumes of atomic groups. These values should prove useful in the analysis of protein packing, protein recognition and ligand design. Our radii for atomic groups were derived from intermolecular distance calculations on a large number (∼30,000) of crystal structures of small organic compounds that contain the same atomic groups to those found in proteins. Our radii show significant differences to previously reported values. We also use this new radii set to determine the packing efficiency in different regions of the protein interior. This analysis shows that, if the surface water molecules are included in the calculations, the overall packing efficiency throughout the protein interior is high and fairly uniform. However, if the water structure is removed, the packing efficiency in peripheral regions of the protein interior is underestimated, by ∼3.5%.}
+}
