### 2025-12-13 maxall41/RustSASA#45 @kliment-olechnovic
Manuscript and documentation issues
Hello, I am one of the reviewers of the RustSASA paper submitted to JOSS - main review thread is https://github.com/openjournals/joss-reviews/issues/9537

Below I would like to point out some issues of the manuscript and documentation that need to be addressed. I will create other GitHub issues for other aspects of the review.

1) Since the Shrake-Rupley was used, its publication must be cited:

Shrake, A; Rupley, JA. (1973). "Environment and exposure to solvent of protein atoms. Lysozyme and insulin". J Mol Biol. 79 (2): 351–71. doi:10.1016/0022-2836(73)90011-9. PMID 4760134.

2) The benchmarking data is not specified in full detail - links or IDs should be provided for the data used. I found in "https://github.com/maxall41/RustSASA/blob/main/paper/eval/get_data.sh" that the data was taken from the AlphaFold database, therefore the AlphaFold database paper should be cited. The used dataset should be clearly identified (via URL or DOI or some other ID) in the paper or at least in the GitHub README, so that the benchmarking experiment could be reproduced (for example, in some future publications).

3) It would be appropriate to cite MDAnalysis, as recommended by the authors of MDAnalysis - https://www.mdanalysis.org/pages/citations/

4) From "https://github.com/maxall41/RustSASA/blob/main/paper/eval/exec_freesasa.sh" it is evident that the author made an appropriate decision to run freesasa using the ''--shrake-rupley -n 100" options for benchmarking, but it would also be prudent to explicitly specify that in the manuscript. I think it is necessary because, as far as I know, the default algorithm in the FreeSASA software is Lee & Richards, and the ''--shrake-rupley -n 100" options override the default FreeSASA behavior.
