### 2025-12-13 maxall41/RustSASA#46 @kliment-olechnovic
Benchmarking-related issues
Hello, I am one of the reviewers of the RustSASA paper submitted to JOSS - main review thread is https://github.com/openjournals/joss-reviews/issues/9537

Below I would like to point out some issues of the benchmarking that need to be addressed. I will create other GitHub issues for other aspects of the review.

1) As I understood, the benchmarking was done by providing the benchmarked tools with the same input coordinates, but not necessarily the same van der Waals radii. The SASA values depend on the radii, and I strongly suspect that the RustSASA vs FreeSASA output differences shown in Figure 1 (the differences are actually quite significant) are most likely the result of differences in how different tools assign radii to atoms. The best solution is to ensure that the radii are the same.

2) The benchmarking uses different parallelization setups for different software. I suggest additionally performing a non-parallelized (single CPU core) benchmark for at least a subset of the data (e.g. the FreeSASA dataset).

3) Related to the previous point, it would be interesting to benchmark a RustSASA version with vectorization (SIMD) disabled - it would provide a clear picture about the importance of SIMD for the performance of RustSASA. I do not know if it could be easily done, therefore it is very much an optional suggestion, but I think it would be cool if the author managed to do it.
