### 2025-12-13 maxall41/RustSASA#47 @kliment-olechnovic
van der Waals radii assignment process is unclear
Hello, I am one of the reviewers of the RustSASA paper submitted to JOSS - main review thread is https://github.com/openjournals/joss-reviews/issues/9537

Below I would like to point out some issues of the input radii assignment process that need to be addressed. I will create other GitHub issues for other aspects of the review.

1) I am confused about how the van der Waals radii assigned. The paper now states that the ProtOr set by Tsai et al. is used. The repo has the "https://github.com/maxall41/RustSASA/blob/main/radii/protor.config" file, but it seems to be not used by the software. As far as I was able to deduce, the radii are assigned by the "pdbtbx" library (it does both the molecular coordinates reading and the radii assignment), and the "pdbtbx" library use the radii from another publication, listed in "https://docs.rs/pdbtbx/0.12.0/pdbtbx/struct.AtomicRadius.html":

Source: Alvarez, S. (2013). A cartography of the van der Waals territories. Dalton Transactions, 42(24), 8617. https://doi.org/10.1039/c3dt50599e

According to "https://github.com/maxall41/pdbtbx/blob/332d29a232755e28c20d12929f9fea72be958e4c/src/structs/elements.rs", the "pdbtbx" library assigns radii based on chemical elements, while the ProtOr dataset also considers specific uniquely named atoms (e.g. different amino acid atoms).

So how does RustSASA assign radii and what is the appropriate citation for the van der Waals radii values?

2) There seems to be no way to provide custom radii for input atoms. Such a feature is crucial for applications where the atom names are not standard (e.g. for coarse-grained molecular simulations).
