### 2025-12-14 9ff2428
- HETATM Filtering by default - Van Der Wals Fallback flag - Updated paper - Improved error handling - Thread customization support - Improved HashMap performance - Added single-threaded benchmark - Added criterion benchmarks
