### 2025-12-14 e8b8a1d
- HETATM Filtering by default - Van Der Wals Fallback flag - Updated paper - Improved error handling - Thread customization support - Improved HashMap performance - Added single-threaded benchmark - Added criterion benchmarks
