+++ paper/paper.bib
@@ -107,3 +107,60 @@ author = {Jerry Tsai and Robin Taylor and Cyrus Chothia and Mark Gerstein},
 keywords = {atomic group radii and volumes, Voronoi polyhedra, packing efficiency, protein interior},
 abstract = {The sizes of atomic groups are a fundamental aspect of protein structure. They are usually expressed in terms of standard sets of radii for atomic groups and of volumes for both these groups and whole residues. Atomic groups, which subsume a heavy-atom and its covalently attached hydrogen atoms into one moiety, are used because the positions of hydrogen atoms in protein structures are generally not known. We have calculated new values for the radii of atomic groups and for the volumes of atomic groups. These values should prove useful in the analysis of protein packing, protein recognition and ligand design. Our radii for atomic groups were derived from intermolecular distance calculations on a large number (∼30,000) of crystal structures of small organic compounds that contain the same atomic groups to those found in proteins. Our radii show significant differences to previously reported values. We also use this new radii set to determine the packing efficiency in different regions of the protein interior. This analysis shows that, if the surface water molecules are included in the calculations, the overall packing efficiency throughout the protein interior is high and fairly uniform. However, if the water structure is removed, the packing efficiency in peripheral regions of the protein interior is underestimated, by ∼3.5%.}
 }
+
+
+@article{ShrakeRupley,
+title = {Environment and exposure to solvent of protein atoms. Lysozyme and insulin},
+journal = {Journal of Molecular Biology},
+volume = {79},
+number = {2},
+pages = {351-371},
+year = {1973},
+issn = {0022-2836},
+doi = {https://doi.org/10.1016/0022-2836(73)90011-9},
+url = {https://www.sciencedirect.com/science/article/pii/0022283673900119},
+author = {A. Shrake and J.A. Rupley},
+abstract = {A computer program is described for calculating the environment and the exposure to solvent of atoms of a protein. The computation is based on the atomic co-ordinates of the protein and on assumptions like those of Lee & Richards (1971). Results for lysozyme and insulin are presented. Changes in exposure to solvent and in the nature of contacts that develop through folding, association reactions and crystallization are described numerically. The computations suggest several generalizations. (a) Lattice contacts within the protein crystal are characterized by a significantly smaller involvement of non-polar side chains and a proportionately greater involvement of ionizable side chains than is found for protein folding or for protein association reactions important for biological function, (b) In helical regions the carbonyl oxygen of the first residue in the helix has high probability of being shielded from solvent, (c) Glycine is among the residues having exposure least affected by folding; this accords with the expectation that it lies at bends of the peptide chain on the surface of the molecule.}
+}
+
+
+@article{AlphaFoldDB,
+    author = {Varadi, Mihaly and Anyango, Stephen and Deshpande, Mandar and Nair, Sreenath and Natassia, Cindy and Yordanova, Galabina and Yuan, David and Stroe, Oana and Wood, Gemma and Laydon, Agata and Žídek, Augustin and Green, Tim and Tunyasuvunakool, Kathryn and Petersen, Stig and Jumper, John and Clancy, Ellen and Green, Richard and Vora, Ankur and Lutfi, Mira and Figurnov, Michael and Cowie, Andrew and Hobbs, Nicole and Kohli, Pushmeet and Kleywegt, Gerard and Birney, Ewan and Hassabis, Demis and Velankar, Sameer},
+    title = {AlphaFold Protein Structure Database: massively expanding the structural coverage of protein-sequence space with high-accuracy models},
[…]
