+++ paper/eval/benchmark.sh
@@ -5,3 +5,7 @@
 hyperfine --warmup 3 --runs 25 -N "../../target/release/rust-sasa ./data/UP000000625_83333_ECOLI_v6/AF-A0A385XJ53-F1-model_v6.pdb ./results/test_out.pdb" --export-json ./results/bench_rustsasa_single.json
 hyperfine --warmup 3 --runs 25 -N "freesasa --shrake-rupley -n 100 --format json ./data/UP000000625_83333_ECOLI_v6/AF-A0A385XJ53-F1-model_v6.pdb" --export-json ./results/bench_freesasa_single.json
 hyperfine --warmup 3 --runs 25 -N "sh exec_biopython.sh ./data/UP000000625_83333_ECOLI_v6/AF-A0A385XJ53-F1-model_v6.pdb ./results/test_out" --export-json ./results/bench_biopython_single.json
+
+hyperfine --warmup 3 --runs 3 "./sasa_batch ./data/UP000000625_83333_ECOLI_v6 ./out/" --export-json ./results/benches/bench_freesasa_singlethread_cpp.json
+hyperfine --warmup 3 --runs 3 "../../target/release/rust-sasa ./data/UP000000625_83333_ECOLI_v6/ ./out/ --format json --threads 1" --export-json ./results/benches/bench_rustsasa_singlethread.json
+hyperfine --warmup 3 --runs 3 "sh scripts/exec_biopython.sh ./data/UP000000625_83333_ECOLI_v6/ ./out/ --no-parallel" --export-json ./results/benches/bench_biopython_singlethread.json
