+++ src/structures/atomic.rs
@@ -19,8 +19,6 @@ pub struct Atom {
     pub id: usize,
     /// Parent Id (8 bytes)
     pub parent_id: Option<isize>,
-    /// Whether this atom is a hydrogen atom (1 byte + padding)
-    pub is_hydrogen: bool,
 }
 
 /// Can be used to specify output resolution of SASA computation for convenience.

[…]
