+++ paper/paper.md
@@ -42,9 +42,9 @@ To evaluate the accuracy of RustSASA calculations, we compared results to Freesa
 
 We evaluated the performance of Freesasa, RustSASA, and Biopython [@biopython] across three evaluations. First, we performed multi-threaded SASA calculations for all proteins in the E. coli proteome. Second, we evaluated the performance of these methods on a single randomly selected protein (A0A385XJ53) from the AlphaFold E. coli proteome. Third, we evaluated the single-threaded performance of RustSASA and Freesasa on the E. coli proteome, Biopython was excluded from this benchmark due to its poor performance hindering timely evaluation.
 
-For the full proteome benchmarks we used Hyperfine [@Hyperfine] with 3 runs and 3 warmup iterations. All methods utilized parallel processing across eight cores. GNU parallel [@Tange2011a] was used to parallelize Freesasa and Biopython, while RustSASA utilized its internal parallelization. RustSASA processed the entire proteome in ~5 seconds compared to ~28 seconds for Freesasa and ~368 seconds for Biopython, representing 5× and 63× speed improvements, respectively. 
+For the full proteome benchmarks (Fig 1A) we used Hyperfine [@Hyperfine] with 3 runs and 3 warmup iterations. All methods utilized parallel processing across eight cores. GNU parallel [@Tange2011a] was used to parallelize Freesasa and Biopython, while RustSASA utilized its internal parallelization. RustSASA processed the entire proteome in ~5 seconds compared to ~28 seconds for Freesasa and ~328 seconds for Biopython, representing 5× and 63× speed improvements, respectively. 
 
-For the single protein benchmark, we used Hyperfine with 3 warmup iterations and 25 runs. RustSASA processed the protein in 4.3ms (±0.5), Freesasa processed the protein in 4.0ms (±0.2), and Biopython processed the protein in 250.8ms (±2.0). In the single threaded benchmark, RustSASA outperforms Freesasa by ~36% demonstrating that RustSASA's performance advantage is not solely due to multi-threading.
+For the single protein benchmark (Fig 1B), we used Hyperfine with 3 warmup iterations and 25 runs. RustSASA processed the protein in 4.3ms (±0.5), Freesasa processed the protein in 4.0ms (±0.2), and Biopython processed the protein in 250.8ms (±2.0). On the single threaded benchmark (Fig 1C), RustSASA processed the proteome in 26.4s compared to 46.2 seconds for Freesasa, representing a ~42% performance improvement, demonstrating that RustSASA's performance advantage is not solely due to multi-threading.
 
 ## Methods
 
@@ -52,7 +52,7 @@ SASA calculation RustSASA computes solvent-accessible surface areas (SASA) using
 
 Atomic radii were assigned according to the ProtOr parameter set introduced by Tsai et al. [@ProtOr]. These radii were applied to all non-hydrogen heavy atoms present in the structures; hydrogen atoms, when present in input files, were ignored for the SASA computations to maintain consistency with common practice for protein SASA estimation. Any non-standard residues or ligands were ignored following the Freesasa methodology [@Mitternacht_2016].
 
[…]
