+++ paper/eval/scripts/exec_rustsasa.sh
@@ -2,5 +2,5 @@
 # $2 = Output
 
 mkdir $2
-# ../../target/aarch64-apple-darwin/release/rust-sasa $1 $2 --format json
-../../target/release/rust-sasa $1 $2 --format json
+../../target/aarch64-apple-darwin/release/rust-sasa $1 $2 --format json
+# ../../target/release/rust-sasa $1 $2 --format json
