### 2025-12-14 c7ab5e5
Increment to v0.7.0 and fix doc issue.
