### 2025-12-14 5fbaa63
Increment to v0.7.0 and fix doc issue.
