### 2025-12-14 f684c15
update doc comments
