### 2025-12-15 maxall41/RustSASA#44 @mittinatten
That's great, I must admit my non-existent rust skills make it bit difficult to follow the code path here, but it sounds like a good solution. I liked that you could read the config file using a macro at compile time.

I'm surprised though that RustSASAs total SASA is still on average larger than FreeSASA. With these changes the results should be almost identical. Maybe the graphs in the readme and paper just aren't updated?
