### 2025-12-15 maxall41/RustSASA#47 @kliment-olechnovic
Thank you for the explanation! Some more questions below.

1) What happens if the residue&atom name combo is not found in the config file? If it is falling back to the default pdbtbx-assigned vdW radius, then it may not be the most flexible solution, because the default radius may not be compatible with the config (e.g. the config has larger radii to account for coarse-graining).

2) What about an option to provide radii for every atom - that is, (x y z r) for every atom?
