### 2025-12-15 maxall41/RustSASA#48 @mittinatten
Non-standard residues
Hi, in the latest version of the paper you state

> Any non-standard residues or ligands were ignored following the freesasa methodology

In FreeSASA the default is to skip HETATM entries, but try to guess the radius of non-standard residues if they are present (and marked as ATOM). They can be included as an option, and you can also specify if atoms for which the radius can not be guessed should be skipped, or cause an error (and if the act of guessing should emit a warning). I don't know if it makes sense to provide all those configuration options in RustSASA, but I think a reasonable default is to include all ATOM entries, and exclude HETATMs, as the latter are usually not part of the protein itself. (I'm sure there's all sorts of variations of this in the PDB dataset, so there will probably be exceptions to any general statement about what to include, and it will depend on what questions you are asking).

It's probably a matter of opinion what is the best default and which configuration options are useful to expose, but it's not correct that the current approach is the same as in FreeSASA. I think sane and well-defined defaults are the most important, since the majority of users probably won't bother customizing anything.

(Tagging JOSS review: https://github.com/openjournals/joss-reviews/issues/9537)
