### 2025-12-15 maxall41/RustSASA#49 @mittinatten
API documentation not quite up to date
The documentation for the API states

> For most users it is recommend that you use calculate_sasa instead.

But that function doesn't exist, and I can see from both the changelog, the code and the example that it's been replaced by the SASAOptions interface. So I guess that's just some left-overs from an older implementation. It's a bit counterintuitive for me that the object that is used to perform the calculations is called _options_, but maybe that's idiomatic in Rust.

Except this one error the docs look adequate to me.

(tagging JOSS review https://github.com/openjournals/joss-reviews/issues/9537)
