### 2025-12-16 55ea6dd
v0.8.0 WIP
