### 2025-12-16 38ff4f4
v0.8.0 WIP
