+++ paper/eval/scripts/quality_scatter.py
@@ -58,23 +58,22 @@ def extract_matching_chain_values(freesasa_chains, rsasa_chains):
     return freesasa_values, rsasa_values
 
 
-def main(freesasa_dir, rsasa_dir, plot_title, output_file):
-    # Directory paths provided via CLI arguments
-
+def process_dataset(freesasa_dir, rsasa_dir):
+    """Process a single dataset and return the data arrays."""
     # Check if directories exist
     if not os.path.exists(freesasa_dir):
         print(f"Directory {freesasa_dir} not found!")
-        return
+        return None, None
     if not os.path.exists(rsasa_dir):
         print(f"Directory {rsasa_dir} not found!")
-        return
+        return None, None
 
     # Get matching files
     matching_files = get_matching_files(freesasa_dir, rsasa_dir)
 
     if not matching_files:
         print("No matching .json files found in both directories!")
-        return
+        return None, None
 
     print(f"Found {len(matching_files)} matching files")
 
@@ -106,7 +105,7 @@ def main(freesasa_dir, rsasa_dir, plot_title, output_file):
 
     if not all_freesasa_values:
         print("No matching chains found!")
-        return
+        return None, None
 
     print(f"Total matching chains: {len(all_freesasa_values)}")
 
@@ -114,20 +113,20 @@ def main(freesasa_dir, rsasa_dir, plot_title, output_file):
     freesasa_values = np.array(all_freesasa_values)
     rsasa_values = np.array(all_rsasa_values)
 
+    return freesasa_values, rsasa_values
+
+
+def create_panel(ax, freesasa_values, rsasa_values, title):
+    """Create a single scatter plot panel."""
     # Calculate Pearson correlation
     correlation, p_value = pearsonr(freesasa_values, rsasa_values)
 
     # Calculate additional statistics
     rmse = np.sqrt(np.mean((freesasa_values - rsasa_values) ** 2))
-    mae = np.mean(np.abs(freesasa_values - rsasa_values))
 
-    plt.figure(figsize=(10, 8))
-    fontsize = 18
-    linewidth = 2.5
-    colors = ["#4477AA"]  # color-blind friendly blue from Paul Tol’s palette
-
-    # Create axes
-    ax = plt.subplot(111)
+    fontsize = 14
+    linewidth = 2.0
+    colors = ["#4477AA"]  # color-blind friendly blue from Paul Tol's palette
 
     ax.spines["right"].set_visible(False)
     ax.spines["top"].set_visible(False)
@@ -151,7 +150,7 @@ def main(freesasa_dir, rsasa_dir, plot_title, output_file):
 
     ax.set_xlabel("FreeSASA Chain Total", fontsize=fontsize + 2)
     ax.set_ylabel("RustSASA Chain Total", fontsize=fontsize + 2)
-    ax.set_title(plot_title, fontsize=fontsize + 4, pad=15)
+    ax.set_title(title, fontsize=fontsize + 2, pad=10)
 
     ax.tick_params(axis="both", which="major", labelsize=fontsize)
     ax.ticklabel_format(style="sci", axis="both", scilimits=(0, 0))
@@ -172,16 +171,63 @@ def main(freesasa_dir, rsasa_dir, plot_title, output_file):
         bbox=dict(boxstyle="round,pad=0.3", facecolor="white", edgecolor="gray", alpha=0.8),
     )
 
-    plt.tight_layout()
+    return correlation, p_value, rmse
+
+
+def main(
[…]
