+++ src/lib.rs
@@ -12,7 +12,6 @@ pub mod options;
 pub use options::{AtomLevel, ChainLevel, ProteinLevel, ResidueLevel, SASAProcessor};
 use utils::consts::ANGLE_INCREMENT;
 pub mod structures;
-mod test;
 pub mod utils;
 
 pub use crate::options::*;
@@ -66,61 +65,21 @@ fn generate_sphere_points(n_points: usize) -> SpherePointsSoA {
 }
 
 #[inline(never)]
-fn precompute_neighbors(
+pub fn precompute_neighbors(
     atoms: &[Atom],
     active_indices: &[usize],
     probe_radius: f32,
     max_radii: f32,
 ) -> Vec<Vec<NeighborData>> {
+    // Same cell_size as original
     let cell_size = probe_radius + max_radii;
-    let grid = SpatialGrid::new(atoms, active_indices, cell_size);
 
-    let mut neighbors = Vec::with_capacity(active_indices.len());
-    let mut temp_candidates = Vec::with_capacity(64);
-    let sr = probe_radius + (max_radii * 2.0);
-    let sr_squared = sr * sr;
+    // Maximum search radius from original: atom.radius + max_radii + 2.0 * probe_radius
+    // Worst case is when atom.radius == max_radii
+    let max_search_radius = max_radii + max_radii + 2.0 * probe_radius;
 
-    for &orig_idx in active_indices {
-        let atom = &atoms[orig_idx];
-        let xyz = atom.position;
-        grid.locate_within_distance([xyz[0], xyz[1], xyz[2]], sr_squared, &mut temp_candidates);
-
-        // Sort candidates by distance (closest first for early exit optimization)
-        let center_pos = atom.position;
-        temp_candidates.sort_unstable_by(|&a_idx, &b_idx| {
-            let pos_a = atoms[a_idx].position;
-            let pos_b = atoms[b_idx].position;
-
-            let dx_a = center_pos[0] - pos_a[0];
-            let dy_a = center_pos[1] - pos_a[1];
-            let dz_a = center_pos[2] - pos_a[2];
-            let dist_sq_a = dx_a * dx_a + dy_a * dy_a + dz_a * dz_a;
-
-            let dx_b = center_pos[0] - pos_b[0];
-            let dy_b = center_pos[1] - pos_b[1];
-            let dz_b = center_pos[2] - pos_b[2];
-            let dist_sq_b = dx_b * dx_b + dy_b * dy_b + dz_b * dz_b;
-
-            dist_sq_a
-                .partial_cmp(&dist_sq_b)
-                .unwrap_or(std::cmp::Ordering::Equal)
-        });
-
-        // Precompute squared thresholds for each neighbor
-        let mut neighbor_data = Vec::with_capacity(temp_candidates.len());
-        for &idx in &temp_candidates {
-            let neighbor = &atoms[idx];
-            let threshold = neighbor.radius + probe_radius;
-            neighbor_data.push(NeighborData {
-                idx: idx as u32,
-                threshold_squared: threshold * threshold,
-            });
-        }
-
-        neighbors.push(neighbor_data);
-    }
-
-    neighbors
+    let grid = SpatialGrid::new(atoms, active_indices, cell_size, max_search_radius);
+    grid.build_all_neighbor_lists(atoms, active_indices, probe_radius, max_radii)
 }
 
 struct AtomSasaKernel<'a> {
@@ -200,29 +159,55 @@ impl<'a> pulp::WithSimd for AtomSasaKernel<'a> {
         }
 
         // Process remainder
+        let mut current_nb = 0;
         for i in 0..sx_rem.len() {
             let sx = sx_rem[i];
             let sy = sy_rem[i];
             let sz = sz_rem[i];
[…]
