+++ tests/common/data.rs
@@ -1,6 +1,6 @@
 #[cfg(test)]
 #[allow(dead_code)]
-const FIXED_LOW_RES_ATOMS: [f32; 2622] = [
+pub const FIXED_LOW_RES_ATOMS: [f32; 2622] = [
     48.24325, 2.5255642, 3.7883465, 12.627821, 16.90931, 20.204514, 50.32727, 66.92745, 2.3533294,
     7.576693, 0.0, 31.569553, 8.454655, 37.883465, 25.255642, 1.1766647, 0.0, 0.0, 17.67895,
     16.90931, 27.781206, 23.99286, 1.1766647, 2.5255642, 0.0, 6.3139105, 0.0, 8.839475, 38.045948,
@@ -238,7 +238,7 @@ const FIXED_LOW_RES_ATOMS: [f32; 2622] = [
 
 #[cfg(test)]
 #[allow(dead_code)]
-const FIXED_HIGH_RES_ATOMS: [f32; 2622] = [
+pub const FIXED_HIGH_RES_ATOMS: [f32; 2622] = [
     48.24325, 2.5255642, 3.7883465, 12.627821, 16.90931, 20.204514, 50.32727, 66.92745, 2.3533294,
     7.576693, 0.0, 31.569553, 8.454655, 37.883465, 25.255642, 1.1766647, 0.0, 0.0, 17.67895,
     16.90931, 27.781206, 23.99286, 1.1766647, 2.5255642, 0.0, 6.3139105, 0.0, 8.839475, 38.045948,
@@ -473,255 +473,3 @@ const FIXED_HIGH_RES_ATOMS: [f32; 2622] = [
     2.5255642, 5.0511284, 1.2627821, 27.477629, 39.146244, 29.591293, 12.943312, 5.0511284,
     10.102257, 18.941732, 34.87545, 11.365039, 12.627821, 52.8416, 46.500607, 47.557438,
 ];
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::options::SASAOptions;
-    use crate::structures::spatial_grid::SpatialGrid;
-    use crate::utils::consts::PROTOR_RADII;
-    use crate::utils::get_protor_radius;
-    use crate::{Atom, AtomLevel, ChainLevel, ProteinLevel, ResidueLevel, calculate_sasa_internal};
-    use approx::assert_abs_diff_eq;
-
-    use std::time::Instant;
-
-    #[test]
-    fn internal_test() {
-        let (pdb, _errors) = pdbtbx::open("./pdbs/example.cif").unwrap();
-        let mut atoms = vec![];
-        for atom in pdb.atoms() {
-            let element = atom.element().unwrap();
-            atoms.push(Atom {
-                position: [
-                    atom.pos().0 as f32,
-                    atom.pos().1 as f32,
-                    atom.pos().2 as f32,
-                ],
-                radius: element.atomic_radius().van_der_waals.unwrap() as f32,
-                id: atom.serial_number(),
-                parent_id: None,
-            })
-        }
-        let start = Instant::now();
-        let sasa = calculate_sasa_internal(&atoms, 1.4, 100, -1);
-        let duration = start.elapsed();
-        // Compare element by element since Vec<f32> doesn't implement AbsDiffEq
-        assert_eq!(sasa.len(), FIXED_LOW_RES_ATOMS.len());
-        for (actual, expected) in sasa.iter().zip(FIXED_LOW_RES_ATOMS.iter()) {
-            assert_abs_diff_eq!(actual, expected, epsilon = 25.0);
-        }
-        println!("Time elapsed (INTERNAL): {duration:?}");
-    }
-
-    #[test]
-    fn external_test() {
[…]
