+++ tests/units.rs
@@ -1,0 +1,277 @@
+mod common;
+
+#[cfg(test)]
+mod tests {
+    use super::common::data::FIXED_LOW_RES_ATOMS;
+    use approx::assert_abs_diff_eq;
+    use rust_sasa::options::SASAOptions;
+    use rust_sasa::structures::spatial_grid::SpatialGrid;
+    use rust_sasa::utils::consts::PROTOR_RADII;
+    use rust_sasa::utils::get_protor_radius;
+    use rust_sasa::{
+        Atom, AtomLevel, ChainLevel, ProteinLevel, ResidueLevel, calculate_sasa_internal,
+    };
+    use std::time::Instant;
+
+    #[test]
+    fn internal_test() {
+        let (pdb, _errors) = pdbtbx::open("./pdbs/example.cif").unwrap();
+        let mut atoms = vec![];
+        for atom in pdb.atoms() {
+            let element = atom.element().unwrap();
+            atoms.push(Atom {
+                position: [
+                    atom.pos().0 as f32,
+                    atom.pos().1 as f32,
+                    atom.pos().2 as f32,
+                ],
+                radius: element.atomic_radius().van_der_waals.unwrap() as f32,
+                id: atom.serial_number(),
+                parent_id: None,
+            })
+        }
+        let start = Instant::now();
+        let sasa = calculate_sasa_internal(&atoms, 1.4, 100, -1);
+        let duration = start.elapsed();
+        // Compare element by element since Vec<f32> doesn't implement AbsDiffEq
+        assert_eq!(sasa.len(), FIXED_LOW_RES_ATOMS.len());
+        for (actual, expected) in sasa.iter().zip(FIXED_LOW_RES_ATOMS.iter()) {
+            assert_abs_diff_eq!(actual, expected, epsilon = 25.0);
+        }
+        println!("Time elapsed (INTERNAL): {duration:?}");
+    }
+
+    #[test]
+    fn external_test() {
+        let (pdb, _errors) = pdbtbx::open("./pdbs/example.cif").unwrap();
+        let protein_sasa = SASAOptions::<ProteinLevel>::new().process(&pdb).unwrap();
+        let chain_sasa = SASAOptions::<ChainLevel>::new().process(&pdb).unwrap();
+
+        let _residue_sasa = SASAOptions::<ResidueLevel>::new().process(&pdb).unwrap();
+
+        let start = Instant::now();
+        let atom_sasa = SASAOptions::<AtomLevel>::new().process(&pdb).unwrap();
+        let duration = start.elapsed();
+        println!("Time elapsed (ATOM): {duration:?}");
+
+        assert_abs_diff_eq!(protein_sasa.global_total, 20268.004, epsilon = 1500.0);
+        // Compare element by element since Vec<f32> doesn't implement AbsDiffEq
+        for (actual, expected) in atom_sasa.iter().zip(FIXED_LOW_RES_ATOMS.iter()) {
+            assert_abs_diff_eq!(actual, expected, epsilon = 25.0);
+        }
+        println!("PROTEIN SASA {protein_sasa:?}");
+        println!("CHAIN SASA {chain_sasa:?}");
+    }
+
+    #[test]
+    fn check_pdb_w_bad_seqadv_record() {
+        let (pdb, _errors) = pdbtbx::open("./pdbs/bad_seqadv_1A06.pdb").unwrap();
+        let protein_sasa = SASAOptions::<ProteinLevel>::new().process(&pdb).unwrap();
+
+        let start = Instant::now();
+        let duration = start.elapsed();
+        println!("Time elapsed (ATOM): {duration:?}");
+
[…]

[…]
