### 2025-12-18 59aad86
snapshot integration tests
