+++ Cargo.toml
@@ -43,6 +43,8 @@ assert_cmd = "2.0.14"
 escargot = "0.5.15"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+quick-xml = { version = "0.38.4", features = ["serialize"] }
+rust-sasa = { path = ".", features = ["cli"] } # Needed for integration tests
 
 [[bench]]
 name = "sasa_benchmarks"
