+++ tests/integration.rs
@@ -15,7 +15,7 @@ mod tests {
         dir.push("out.cif");
         let mut cmd = Command::new(cargo_bin!("rust-sasa"));
         let cmd = cmd
-            .arg("./pdbs/example.cif")
+            .arg("./tests/data/pdbs/example.cif")
             .arg(dir.clone().into_os_string().into_string().unwrap())
             .arg("--output-depth")
             .arg("atom");
@@ -36,7 +36,7 @@ mod tests {
         dir.push("out.pdb");
         let mut cmd = Command::new(cargo_bin!("rust-sasa"));
         let cmd = cmd
-            .arg("./pdbs/example.cif")
+            .arg("./tests/data/pdbs/example.cif")
             .arg(dir.clone().into_os_string().into_string().unwrap())
             .arg("--output-depth")
             .arg("atom");
@@ -58,7 +58,7 @@ mod tests {
         dir.push("out.json");
         let mut cmd = Command::new(cargo_bin!("rust-sasa"));
         let cmd = cmd
-            .arg("./pdbs/example.cif")
+            .arg("./tests/data/pdbs/example.cif")
             .arg(dir.clone().into_os_string().into_string().unwrap())
             .arg("--output-depth")
             .arg("atom");
@@ -83,7 +83,7 @@ mod tests {
         dir.push("out.xml");
         let mut cmd = Command::new(cargo_bin!("rust-sasa"));
         let cmd = cmd
-            .arg("./pdbs/example.cif")
+            .arg("./tests/data/pdbs/example.cif")
             .arg(dir.clone().into_os_string().into_string().unwrap())
             .arg("--output-depth")
             .arg("atom");
@@ -109,7 +109,7 @@ mod tests {
         dir.push("out.xml");
         let mut cmd = Command::new(cargo_bin!("rust-sasa"));
         let cmd = cmd
-            .arg("./pdbs/example.cif")
+            .arg("./tests/data/pdbs/example.cif")
             .arg(dir.clone().into_os_string().into_string().unwrap())
             .arg("--probe-radius")
             .arg("1.5")
@@ -130,7 +130,7 @@ mod tests {
         dir.push("out.xml");
         let mut cmd = Command::new(cargo_bin!("rust-sasa"));
         let cmd = cmd
-            .arg("./pdbs/example.cif")
+            .arg("./tests/data/pdbs/example.cif")
             .arg(dir.clone().into_os_string().into_string().unwrap())
             .arg("--probe-radius")
             .arg("1.5")
@@ -151,7 +151,7 @@ mod tests {
         dir.push("out.xml");
         let mut cmd = Command::new(cargo_bin!("rust-sasa"));
         let cmd = cmd
-            .arg("./pdbs/example.cif")
+            .arg("./tests/data/pdbs/example.cif")
             .arg(dir.clone().into_os_string().into_string().unwrap())
             .arg("--threads")
             .arg("1")
@@ -169,7 +169,7 @@ mod tests {
         dir.push("out.xml");
         let mut cmd = Command::new(cargo_bin!("rust-sasa"));
         let cmd = cmd
-            .arg("./pdbs/example.cif")
+            .arg("./tests/data/pdbs/example.cif")
             .arg(dir.clone().into_os_string().into_string().unwrap())
             .arg("--output-depth")
             .arg("atom");
@@ -183,7 +183,7 @@ mod tests {
         let dir = env::temp_dir();
         let mut cmd = Command::new(cargo_bin!("rust-sasa"));
         let cmd = cmd
-            .arg("./pdbs/")
+            .arg("./tests/data/pdbs/")
             .arg(dir.clone().into_os_string().into_string().unwrap())
             .arg("--format")
[…]
