### 2025-12-20 801108c
molecular dynamics simulations
