+++ paper/paper.bib
@@ -164,3 +164,65 @@ eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/jcc.21787},
 abstract = {Abstract MDAnalysis is an object-oriented library for structural and temporal analysis of molecular dynamics (MD) simulation trajectories and individual protein structures. It is written in the Python language with some performance-critical code in C. It uses the powerful NumPy package to expose trajectory data as fast and efficient NumPy arrays. It has been tested on systems of millions of particles. Many common file formats of simulation packages including CHARMM, Gromacs, Amber, and NAMD and the Protein Data Bank format can be read and written. Atoms can be selected with a syntax similar to CHARMM's powerful selection commands. MDAnalysis enables both novice and experienced programmers to rapidly write their own analytical tools and access data stored in trajectories in an easily accessible manner that facilitates interactive explorative analysis. MDAnalysis has been tested on and works for most Unix-based platforms such as Linux and Mac OS X. It is freely available under the GNU General Public License from http://mdanalysis.googlecode.com. © 2011 Wiley Periodicals, Inc. J Comput Chem 2011},
 year = {2011}
 }
+
+@Article{Espigares2020,
+author={Rodr{\'i}guez-Espigares, Ismael
+and Torrens-Fontanals, Mariona
+and Tiemann, Johanna K. S.
+and Aranda-Garc{\'i}a, David
+and Ram{\'i}rez-Anguita, Juan Manuel
+and Stepniewski, Tomasz Maciej
+and Worp, Nathalie
+and Varela-Rial, Alejandro
+and Morales-Pastor, Adri{\'a}n
+and Medel-Lacruz, Brian
+and P{\'a}ndy-Szekeres, G{\'a}sp{\'a}r
+and Mayol, Eduardo
+and Giorgino, Toni
+and Carlsson, Jens
+and Deupi, Xavier
+and Filipek, Slawomir
+and Filizola, Marta
+and G{\'o}mez-Tamayo, Jos{\'e} Carlos
+and Gonzalez, Angel
+and Guti{\'e}rrez-de-Ter{\'a}n, Hugo
+and Jim{\'e}nez-Ros{\'e}s, Mireia
+and Jespers, Willem
+and Kapla, Jon
+and Khelashvili, George
+and Kolb, Peter
+and Latek, Dorota
+and Marti-Solano, Maria
+and Matricon, Pierre
+and Matsoukas, Minos-Timotheos
+and Miszta, Przemyslaw
+and Olivella, Mireia
+and Perez-Benito, Laura
+and Provasi, Davide
+and R{\'i}os, Santiago
+and R. Torrecillas, Iv{\'a}n
+and Sallander, Jessica
+and Sztyler, Agnieszka
+and Vasile, Silvana
+and Weinstein, Harel
+and Zachariae, Ulrich
+and Hildebrand, Peter W.
+and De Fabritiis, Gianni
+and Sanz, Ferran
+and Gloriam, David E.
+and Cordomi, Arnau
+and Guix{\`a}-Gonz{\'a}lez, Ramon
+and Selent, Jana},
+title={GPCRmd uncovers the dynamics of the 3D-GPCRome},
+journal={Nature Methods},
+year={2020},
+month={Aug},
+day={01},
+volume={17},
+number={8},
+pages={777-787},
[…]
