### 2025-12-20 88b6607
molecular dynamics simulations
