+++ paper/paper.md
@@ -17,19 +17,13 @@ bibliography: paper.bib
 
 ---
 
-TODO: Include molecular dynamics results in performance section and update summary.
-
-TODO: Implement all TODOs.
-
 # Summary
 
-Solvent accessible surface area (SASA) calculations are fundamental for understanding protein structure, function, and dynamics in computational biology. These calculations quantify the surface area of biomolecules accessible to solvent molecules, providing insights into protein folding, stability, and intermolecular interactions. The Shrake-Rupley algorithm has served as the standard for SASA calculations since 1973, but existing implementations often become computational bottlenecks when analyzing large protein datasets. As proteomics datasets continue to grow with initiatives like AlphaFold producing hundreds of millions of predicted protein structures, the need for efficient SASA calculation tools has increased dramatically. RustSASA addresses this challenge by providing a high-performance implementation of the Shrake-Rupley algorithm written in pure Rust, delivering a 5× speed improvement over FreeSASA while maintaining calculation accuracy and providing interfaces for multiple programming languages and frameworks (i.e. MDAnalysis). The source code is freely available at https://github.com/maxall41/RustSASA.
+Solvent accessible surface area (SASA) calculations are fundamental for understanding protein structure, function, and dynamics in computational biology, providing insights into protein folding, stability, and intermolecular interactions. The Shrake-Rupley algorithm has served as the standard method since 1973, but existing implementations often become computational bottlenecks when analyzing large protein datasets or molecular dynamics trajectories. As structural biology databases continue to expand, exemplified by AlphaFold's hundreds of millions of predicted structures, the need for efficient SASA calculation tools has increased dramatically. RustSASA addresses this challenge through a high-performance implementation written in Rust, leveraging efficient parallelization abstractions and SIMD instructions. Benchmarking demonstrates a 5× speed improvement over FreeSASA for proteome-scale calculations and a 20× improvement for molecular dynamics trajectory analysis. Validation against FreeSASA yields Pearson correlation coefficients exceeding 0.99 on both the predicted *E. coli* proteome and the FreeSASA evaluation dataset, confirming calculation accuracy. RustSASA provides interfaces for multiple programming languages (Rust and Python), a command-line interface, and integration with the MDAnalysis framework, ensuring broad accessibility across the computational biology community. The source code is freely available at https://github.com/maxall41/RustSASA.
 
 # Statement of need
 
-Current SASA calculation tools represent a significant computational bottleneck in structural biology workflows, particularly for molecular dynamics simulations and high-throughput analyses. Popular implementations such as those in Biopython and FreeSASA, while accurate, become prohibitively slow when processing large protein datasets. RustSASA addresses this performance gap by leveraging Rust's efficient parallelization abstractions (via Rayon) and readily available SIMD instructions (via Pulp). These optimizations enable RustSASA's performance advantage over the simpler C implementation of the same algorithm in FreeSASA.
-
-Benchmarking on representative protein structures demonstrates that RustSASA achieves a 5× improvement over FreeSASA, and a 63× performance improvement over Biopython. This performance advantage reduces computational costs for high-throughput structural analyses and makes large-scale comparative studies feasible. Furthermore, RustSASA's multi-language support (Rust and Python), command-line interface, and MDAnalysis package [@gowers2016; @Agrawal2011] ensure broad accessibility across the computational biology community.
[…]
