+++ paper/paper.md
@@ -19,11 +19,11 @@ bibliography: paper.bib
 
 # Summary
 
-Solvent accessible surface area (SASA) calculations are fundamental for understanding protein structure, function, and dynamics in computational biology, providing insights into protein folding, stability, and intermolecular interactions. The Shrake-Rupley algorithm has served as the standard method since 1973, but existing implementations often become computational bottlenecks when analyzing large protein datasets or molecular dynamics trajectories. As structural biology databases continue to expand, exemplified by AlphaFold's hundreds of millions of predicted structures, the need for efficient SASA calculation tools has increased dramatically. RustSASA addresses this challenge through a high-performance implementation written in Rust, leveraging efficient parallelization abstractions and SIMD instructions. Benchmarking demonstrates a 5× speed improvement over FreeSASA for proteome-scale calculations and a 20× improvement for molecular dynamics trajectory analysis. Validation against FreeSASA yields Pearson correlation coefficients exceeding 0.99 on both the predicted *E. coli* proteome and the FreeSASA evaluation dataset, confirming calculation accuracy. RustSASA provides interfaces for multiple programming languages (Rust and Python), a command-line interface, and integration with the MDAnalysis framework, ensuring broad accessibility across the computational biology community. The source code is freely available at https://github.com/maxall41/RustSASA.
+Solvent accessible surface area (SASA) calculations are fundamental for understanding protein structure, function, and dynamics in computational biology, providing insights into protein folding, stability, and intermolecular interactions. The Shrake-Rupley algorithm has served as the standard method since 1973, but existing implementations often become computational bottlenecks when analyzing large protein datasets or molecular dynamics trajectories. As structural biology databases continue to expand, exemplified by AlphaFold's hundreds of millions of predicted structures, the need for efficient SASA calculation tools has increased dramatically. RustSASA addresses this challenge through a high-performance implementation written in Rust, leveraging efficient parallelization abstractions and SIMD instructions. Benchmarking demonstrates a 5× speed improvement over FreeSASA for proteome-scale calculations and a 20× improvement for molecular dynamics trajectory analysis. Validation against FreeSASA yields Pearson correlation coefficients exceeding 0.99 on both the predicted *E. coli* proteome and the FreeSASA evaluation dataset, confirming calculation accuracy. RustSASA provides interfaces for multiple programming languages (Rust and Python), a command-line interface, and integration with the MDAnalysis framework [@gowers2016; @Agrawal2011], ensuring broad accessibility across the computational biology community. The source code is freely available at https://github.com/maxall41/RustSASA.
 
 # Statement of need
 
-As proteomics datasets continue to grow with initiatives like AlphaFold producing hundreds of millions of predicted protein structures, existing SASA calculation tools have become a significant bottleneck in structural biology workflows. Popular implementations such as those in Biopython and FreeSASA, while accurate, become prohibitively slow when processing large protein datasets or extended molecular dynamics trajectories. RustSASA addresses this performance gap by leveraging Rust's efficient parallelization abstractions via Rayon and readily available SIMD instructions via Pulp. These optimizations enable RustSASA's performance advantage over the simpler C implementation of the same algorithm in FreeSASA. The resulting performance gains reduce computational costs for high-throughput structural analyses and make previously intractable large-scale comparative studies feasible.
[…]
