### 2025-12-20 b3a618b
add copyright notice to files
