+++ tests/common/mod.rs
@@ -1,2 +1,3 @@
+// Copyright (c) 2024 Maxwell Campbell. Licensed under the MIT License.
 pub mod data;
 pub mod io;
