+++ src/structures/atomic.rs
@@ -1,3 +1,4 @@
+// Copyright (c) 2024 Maxwell Campbell. Licensed under the MIT License.
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 
