### 2025-12-20 fe12b98
add copyright notice to files
