+++ tests/common/io.rs
@@ -1,3 +1,4 @@
+// Copyright (c) 2024 Maxwell Campbell. Licensed under the MIT License.
 use rust_sasa::SASAResult;
 use serde::Deserialize;
 use std::fs;
