+++ src/structures/spatial_grid.rs
@@ -1,3 +1,4 @@
+// Copyright (c) 2024 Maxwell Campbell. Licensed under the MIT License.
 use crate::{Atom, NeighborData};
 
 pub struct SpatialGrid {
