+++ CHANGELOG.md
@@ -4,11 +4,12 @@
 
 * Removed `nalgebra` dependancy.
 * Updated `criterion` to version 0.7.0.
-* Paper updated
+* Paper updated.
 * CLI dependencies gated behind `cli` feature to make library more lightweight. Library dependencies removed: clap, serde, serde_json, quick-xml, indicatif, and mimalloc.
 * Improved SASA calculation accuracy by expanding neighbor search radius.
 * Improved performance of SpatialGrid data structure (using Verlet lists) to compensate for performance penalty of increased neighbor search radius.
 * Improved test suite.
+* Added copyright notice to all files.
 
 ## Version 0.7.0 
 
