### 2025-12-20 e30e5e6
fix pgo CLI build
