### 2025-12-20 03b42c9
fix pgo build
