### 2025-12-20 bc8ba03
fix pgo build
