### 2025-12-20 d827984
fix alt loc selection
