### 2025-12-20 ac5348e
read_radii_from_occupancy support
