+++ src/options.rs
@@ -65,6 +65,7 @@ pub struct SASAOptions<T> {
     radii_config: Option<FnvHashMap<String, FnvHashMap<String, f32>>>,
     allow_vdw_fallback: bool,
     include_hetatms: bool,
+    read_radii_from_occupancy: bool,
     _marker: PhantomData<T>,
 }
 
@@ -78,21 +79,25 @@ pub type AtomsMappingResult = Result<(Vec<Atom>, FnvHashMap<isize, Vec<usize>>),
 
 /// Macro to reduce duplication in atom building logic
 macro_rules! build_atom {
-    ($atoms:expr, $atom:expr, $element:expr, $residue_name:expr, $atom_name:expr, $parent_id:expr, $radii_config:expr, $allow_vdw_fallback:expr, $id:expr) => {{
-        let radius = match get_radius($residue_name, $atom_name, $radii_config) {
-            Some(r) => r,
-            None => {
-                if $allow_vdw_fallback {
-                    $element
-                        .atomic_radius()
-                        .van_der_waals
-                        .context(VanDerWaalsMissingSnafu)? as f32
-                } else {
-                    return Err(SASACalcError::RadiusMissing {
-                        residue_name: $residue_name.to_string(),
-                        atom_name: $atom_name.to_string(),
-                        element: $element.to_string(),
-                    });
+    ($atoms:expr, $atom:expr, $element:expr, $residue_name:expr, $atom_name:expr, $parent_id:expr, $radii_config:expr, $allow_vdw_fallback:expr, $read_radii_from_occupancy:expr, $id:expr) => {{
+        let radius = if $read_radii_from_occupancy {
+            $atom.occupancy() as f32
+        } else {
+            match get_radius($residue_name, $atom_name, $radii_config) {
+                Some(r) => r,
+                None => {
+                    if $allow_vdw_fallback {
+                        $element
+                            .atomic_radius()
+                            .van_der_waals
+                            .context(VanDerWaalsMissingSnafu)? as f32
+                    } else {
+                        return Err(SASACalcError::RadiusMissing {
+                            residue_name: $residue_name.to_string(),
+                            atom_name: $atom_name.to_string(),
+                            element: $element.to_string(),
+                        });
+                    }
                 }
             }
         };
@@ -127,6 +132,7 @@ pub trait SASAProcessor {
         allow_vdw_fallback: bool,
         include_hydrogens: bool,
         include_hetatms: bool,
+        read_radii_from_occupancy: bool,
     ) -> AtomsMappingResult;
 }
 
@@ -148,6 +154,7 @@ impl SASAProcessor for AtomLevel {
         allow_vdw_fallback: bool,
         include_hydrogens: bool,
         include_hetatms: bool,
+        read_radii_from_occupancy: bool,
     ) -> Result<(Vec<Atom>, FnvHashMap<isize, Vec<usize>>), SASACalcError> {
         let mut atoms = vec![];
         for residue in pdb.residues() {
@@ -172,6 +179,7 @@ impl SASAProcessor for AtomLevel {
                         None,
                         radii_config,
                         allow_vdw_fallback,
+                        read_radii_from_occupancy,
                         combine_hash(conformer_alt, atom.serial_number())
                     );
                 }
@@ -223,6 +231,7 @@ impl SASAProcessor for ResidueLevel {
         allow_vdw_fallback: bool,
         include_hydrogens: bool,
         include_hetatms: bool,
+        read_radii_from_occupancy: bool,
     ) -> Result<(Vec<Atom>, FnvHashMap<isize, Vec<usize>>), SASACalcError> {
         let mut atoms = vec![];
         let mut parent_to_atoms = FnvHashMap::default();
[…]
