### 2025-12-20 0b351ee
expand quality test suite
