+++ tests/quality.rs
@@ -16,20 +16,41 @@ mod tests {
 
     const RMSE_BASELINE: f64 = 614.26; //RustSASA RMSE as of v0.8.0
     const TOLERANCE: f64 = 0.01; // 1%
-    const MAX_RMSE: f64 = RMSE_BASELINE * (1.0 + TOLERANCE);
 
-    /// Load FreeSASA reference data and extract chain totals
+    /// Load FreeSASA reference data and extract chain totals or file-level total
+    /// For use_file_total=true, sums all chains and returns a single value keyed by filename
+    /// For use_file_total=false, returns individual chain totals keyed by chain label
     fn load_freesasa_chains(
         file_path: &Path,
+        use_file_total: bool,
     ) -> Result<HashMap<String, f64>, Box<dyn std::error::Error>> {
         let content = fs::read_to_string(file_path)?;
         let data: FreeSASAOutput = serde_json::from_str(&content)?;
 
         let mut chain_totals = HashMap::new();
-        for result in data.results {
-            for structure in result.structure {
-                for chain in structure.chains {
-                    chain_totals.insert(chain.label, chain.area.total);
+
+        if use_file_total {
+            // Sum all chains for file-level comparison (atom/protein depths)
+            let mut total = 0.0;
+            for result in data.results {
+                for structure in result.structure {
+                    for chain in structure.chains {
+                        total += chain.area.total;
+                    }
+                }
+            }
+            let filename = file_path
+                .file_stem()
+                .and_then(|s| s.to_str())
+                .unwrap_or("unknown");
+            chain_totals.insert(filename.to_string(), total);
+        } else {
+            // Extract individual chain totals (residue/chain depths)
+            for result in data.results {
+                for structure in result.structure {
+                    for chain in structure.chains {
+                        chain_totals.insert(chain.label, chain.area.total);
+                    }
                 }
             }
         }
@@ -37,7 +58,8 @@ mod tests {
         Ok(chain_totals)
     }
 
-    /// Load RustSASA output and aggregate by chain
+    /// Load RustSASA output and aggregate for comparison (handles all output depths)
+    /// Returns either chain-level totals or file-level total depending on output depth
     fn load_rustsasa_chains(
         file_path: &Path,
     ) -> Result<HashMap<String, f64>, Box<dyn std::error::Error>> {
@@ -45,12 +67,37 @@ mod tests {
         let data: SASAResult = serde_json::from_str(&content)?;
 
         let mut chain_totals: HashMap<String, f64> = HashMap::new();
-        if let SASAResult::Residue(residues) = data {
-            for residue in residues {
-                *chain_totals.entry(residue.chain_id).or_insert(0.0) += residue.value as f64;
+
+        match data {
+            SASAResult::Atom(atoms) => {
+                // Atom level has no chain information, so we sum all atoms
+                // and use filename as key for comparison
+                let total: f64 = atoms.iter().map(|&v| v as f64).sum();
+                let filename = file_path
+                    .file_stem()
+                    .and_then(|s| s.to_str())
+                    .unwrap_or("unknown");
+                chain_totals.insert(filename.to_string(), total);
             }
-        } else {
-            return Err("Expected Residue level SASAResult".into());
+            SASAResult::Residue(residues) => {
+                for residue in residues {
+                    *chain_totals.entry(residue.chain_id.clone()).or_insert(0.0) +=
+                        residue.value as f64;
+                }
+            }
+            SASAResult::Chain(chains) => {
+                for chain in chains {
[…]
