### 2025-12-21 ccbad44
test radii in occupancy
