+++ tests/quality.rs
@@ -256,4 +256,192 @@ mod tests {
     fn test_quality_protein_depth() {
         run_quality_test("protein");
     }
+
+    /// Copy PDB files and insert protor radii values into the occupancy column
+    fn prepare_pdbs_with_radii_in_occupancy(
+        input_dir: &Path,
+        output_dir: &Path,
+    ) -> Result<(), Box<dyn std::error::Error>> {
+        use pdbtbx::{ReadOptions, StrictnessLevel};
+        use rust_sasa::utils::get_protor_radius;
+
+        fs::create_dir_all(output_dir)?;
+
+        // Process each PDB file in the input directory
+        for entry in fs::read_dir(input_dir)? {
+            let entry = entry?;
+            let path = entry.path();
+
+            if !path.is_file() {
+                continue;
+            }
+
+            let path_str = path
+                .to_str()
+                .ok_or_else(|| format!("Invalid UTF-8 in path: {}", path.display()))?;
+
+            // Read the PDB file
+            let (mut pdb, _errors) = ReadOptions::default()
+                .set_level(StrictnessLevel::Loose)
+                .read(path_str)
+                .map_err(|e| format!("Failed to read PDB {}: {:?}", path.display(), e))?;
+
+            // Modify occupancy values to contain protor radii
+            for residue in pdb.residues_mut() {
+                let Some(residue_name) = residue.name().map(|s| s.to_string()) else {
+                    continue;
+                };
+
+                for conformer in residue.conformers_mut() {
+                    for atom in conformer.atoms_mut() {
+                        let atom_name = atom.name();
+
+                        // Look up the radius using get_protor_radius with fallback to VdW radii
+                        let radius = if let Some(r) = get_protor_radius(&residue_name, atom_name) {
+                            r
+                        } else {
+                            let element = atom.element().ok_or_else(|| {
+                                format!(
+                                    "Element missing for atom '{}' in residue '{}' in file {}",
+                                    atom_name,
+                                    residue_name,
+                                    path.display()
+                                )
+                            })?;
+                            element.atomic_radius().van_der_waals.ok_or_else(|| {
+                                format!(
+                                    "Van der Waals radius missing for element '{}' in file {}",
+                                    element,
+                                    path.display()
+                                )
+                            })? as f32
+                        };
+
+                        // Set occupancy to the radius value
+                        atom.set_occupancy(radius as f64)
+                            .expect("NEVER FAIL: New occupancy is always finite");
+                    }
+                }
+            }
+
+            // Save the modified PDB to the output directory
+            let output_path = output_dir.join(entry.file_name());
+            let output_path_str = output_path.to_str().ok_or_else(|| {
+                format!("Invalid UTF-8 in output path: {}", output_path.display())
+            })?;
+            pdbtbx::save(&pdb, output_path_str, StrictnessLevel::Loose)
+                .map_err(|e| format!("Failed to save PDB to {}: {:?}", output_path.display(), e))?;
+        }
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_quality_read_radii_from_occupancy() {
+        // Setup directories
+        let reference_dir = PathBuf::from("./tests/data/freesasa_reference/");
+        let pdb_dir = PathBuf::from("./tests/data/freesasa_pdbs/");
+        let modified_pdb_dir = env::temp_dir().join("rustsasa_quality_test_occupancy_pdbs");
[…]
