### 2025-12-22 fb131be
fix hetatm chain assignment issue
