### 2025-12-23 7e554d0
fix pgo build
