### 2025-12-23 maxall41/RustSASA#46 @maxall41
I have fixed this issue in v0.9.0 of RustSASA; it was due to atom SASA values not being correctly mapped onto residues in multi-chain proteins. As of v0.9.0, RustSASA takes the first altLoc.
