### 2025-12-23 maxall41/RustSASA#47 @maxall41
As of v0.9.0 of RustSASA, users can specify radii values in the occupancy column of inputted proteins.
