### 2025-12-23 maxall41/RustSASA#47 @kliment-olechnovic
Thanks, it is a useful feature.
