### 2025-12-23 maxall41/RustSASA#46 @kliment-olechnovic
Thanks, it is a crucially important fix. Now the whole Figure 1 looks as expected.
