### 2025-12-26 #9537 @prashjha
Hi @maxall41, I created a pull request with minor edits to your JOSS paper: maxall41/RustSASA#57. Also, please send me the details mentioned in [my earlier comment](https://github.com/openjournals/joss-reviews/issues/9537#issuecomment-3692820370).
