### 2025-12-26 #9537 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
