### 2026-01-07 cc43c8b
Fix paper issues
