+++ paper/paper.bib
@@ -84,7 +84,7 @@ url={https://doi.org/10.1038/s41586-021-03819-2}
   title = {GNU Parallel - The Command-Line Power Tool},
   author = {O. Tange},
   address = {Frederiksberg, Denmark},
-  journal = {;login: The USENIX Magazine},
+  journal = {. ;login: The USENIX Magazine},
   month = {Feb},
   number = {1},
   volume = {36},
@@ -94,7 +94,7 @@ url={https://doi.org/10.1038/s41586-021-03819-2}
 }
 
 @article{ProtOr,
-title = {The packing density in proteins: standard radii and volumes11Edited by J. M. Thornton},
+title = {The packing density in proteins: standard radii and volumes},
 journal = {Journal of Molecular Biology},
 volume = {290},
 number = {1},
@@ -143,7 +143,7 @@ abstract = {A computer program is described for calculating the environment and 
 
 @article{gowers2016,
   author = {Gowers, Richard J. and Linke, Max and Barnoud, Jonathan and Reddy, Tyler J. E. and Melo, Manuel N. and Seyler, Sean L. and Domański, Jan and Dotson, David L. and Buchoux, Sébastien and Kenney, Ian M. and Beckstein, Oliver},
-  title = {MDAnalysis: A Python Package for the Rapid Analysis of Molecular Dynamics Simulations},
+  title = {MDAnalysis: A {Python} Package for the Rapid Analysis of Molecular Dynamics Simulations},
   journal = {SciPy 2016},
   year = {2016},
   doi = {10.25080/Majora-629e541a-00e},
