### 2026-01-07 660ebc5
refine wording in paper
