### 2026-01-07 #9537 @maxall41
@rkurchin I have corrected these issues in the latest draft of the manuscript.
