### 2026-01-12 #9537 @rkurchin
Congrats to @maxall41 and thanks to editor @prashjha and reviewers @mittinatten and @kliment-olechnovic, JOSS's work couldn't happen without you! 🎉
